### 2017-11-30 #479 @Kevin-Mattheus-Moerman
@cMadan I recommend the following as potential reviewers to this [JOSS](http://joss.theoj.org/about) submission: @gpeyre @siboles
