### 2017-11-30 #479 @cMadan
@Kevin-Mattheus-Moerman - it's great to see this project submitted to JOSS! I look forward to overseeing the review process :)

@gpeyre @siboles - are either of you available to review this submission? Thanks!
