### 2017-12-01 #479 @siboles
Yes, I am happy to review this. I am a bit under load at the moment, but
will get to it before the holidays.

Best,

Scott
