### 2017-12-01 #479 @gpeyre
Hello,

I am sorry but I will not be able to make the review.

Best

Gabriel

> Le 30 nov. 2017 à 23:22, Kevin Mattheus Moerman <notifications@github.com> a écrit :
>
> @cMadan I recommend the following as potential reviewers to this JOSS submission: @gpeyre @siboles
>
