### 2017-12-06 72ff963
Added patchNanFix
