+++ lib/plotV.m
@@ -1,44 +1,44 @@
 function varargout=plotV(V,varargin)
 
-nDims=size(V,2);
-
-%Ad zeros if input is 2D
-if nDims==2;
-    V(:,3)=0; 
+if ~isempty(V)
+    nDims=size(V,2);
+    
+    %Ad zeros if input is 2D
+    if nDims==2
+        V(:,3)=0;
+    end
+    
+    hp=plot3(V(:,1),V(:,2),V(:,3),varargin{:});
+    
+else
+    hp=[];
 end
 
-hp=plot3(V(:,1),V(:,2),V(:,3),varargin{:});
-
-% varargout{1}=hp;
 switch nargout
     case 1
         varargout{1}=hp;
 end
-   
 
-
-
- 
-%% 
-% _*GIBBON footer text*_ 
-% 
+%%
+% _*GIBBON footer text*_
+%
 % License: <https://github.com/gibbonCode/GIBBON/blob/master/LICENSE>
-% 
+%
 % GIBBON: The Geometry and Image-based Bioengineering add-On. A toolbox for
 % image segmentation, image-based modeling, meshing, and finite element
 % analysis.
-% 
+%
 % Copyright (C) 2017  Kevin Mattheus Moerman
-% 
+%
 % This program is free software: you can redistribute it and/or modify
 % it under the terms of the GNU General Public License as published by
 % the Free Software Foundation, either version 3 of the License, or
 % (at your option) any later version.
-% 
+%
 % This program is distributed in the hope that it will be useful,
 % but WITHOUT ANY WARRANTY; without even the implied warranty of
 % MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 % GNU General Public License for more details.
-% 
+%
 % You should have received a copy of the GNU General Public License
 % along with this program.  If not, see <http://www.gnu.org/licenses/>.
