+++ data/temp/tempModel.txt
@@ -36,17 +36,17 @@
 
  FILES USED
 ===========================================================================
-	Input file : D:\MATLAB\GIT\GIBBON\data\temp\tempModel.feb
-	Plot file  : D:\MATLAB\GIT\GIBBON\data\temp\tempModel.xplt
-	Log file   : D:\MATLAB\GIT\GIBBON\data\temp\tempModel.txt
+	Input file : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel.feb
+	Plot file  : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel.xplt
+	Log file   : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel.txt
 
 
  MESH INFO
 ===========================================================================
 	Number of materials ............................ : 1
 	Number of domains .............................. : 1
-	Number of nodes ................................ : 6832
-	Number of solid elements ....................... : 3584
+	Number of nodes ................................ : 2462
+	Number of solid elements ....................... : 13074
 
 
  CONTROL DATA
@@ -54,14 +54,14 @@
 	Module type .................................... : solid
 	Analysis type .................................. : quasi-static
 	Plane strain mode .............................. : no
-	Number of timesteps ............................ : 15
-	Time step size ................................. : 0.066667
+	Number of timesteps ............................ : 10
+	Time step size ................................. : 0.1
 	Auto time stepper activated .................... : yes
-	  Optimal nr of iterations ..................... : 5
-	  Minimum allowable step size .................. : 0.000667
-	  Maximum allowable step size .................. : 0.066667
+	  Optimal nr of iterations ..................... : 10
+	  Minimum allowable step size .................. : 0.001
+	  Maximum allowable step size .................. : 0.1
 	Number of loadcurves ........................... : 1
-	Number of displacement boundary conditions ..... : 3
+	Number of displacement boundary conditions ..... : 0
 	Number of concentrated nodal forces ............ : 0
 	Max nr of stiffness reformations ............... : 25
 	per time steps
@@ -70,11 +70,11 @@
 	Linesearch convergence tolerance ............... : 0.9
 	Minimum line search size ....................... : 0.01
 	Maximum number of line search iterations ....... : 5
-	Max condition number ........................... : 4.94066e-319
+	Max condition number ........................... : 0.01
 	Displacement convergence tolerance ............. : 0.001
 	Energy convergence tolerance ................... : 0.01
 	Residual convergence tolerance ................. : 0
-	Minimal residual value ......................... : 1e-020
+	Minimal residual value ......................... : 1e-20
 
 
  OUTPUT DATA
@@ -98,13 +98,13 @@
 	minaug ............................................ : 0
 	maxaug ............................................ : 0
 	c1 ................................................ : 0.001
-	c2 ................................................ : 0.001
+	c2 ................................................ : 0
 	c3 ................................................ : 0
 	c4 ................................................ : 0
 	c5 ................................................ : 0
 	c6 ................................................ : 0
-	m1 ................................................ : 2
-	m2 ................................................ : -2
+	m1 ................................................ : 12
+	m2 ................................................ : 1
[…]
