### 2017-12-13 #479 @cMadan
@siboles, I wasn't sure if someone would be more available to review, but I suppose it has been two weeks now. I'm happy to assign you as the reviewer and get things started.

@wspr, if you're available to contribute to the review, please feel free to also evaluate the submission and make comments!
