### 2017-12-14 0e0aef6
Added test, version statements to readme, added multi-step tension, compression, shear demo
