+++ lib/gwaitbar.m
@@ -1,0 +1,287 @@
+function fout = gwaitbar(x,whichbar, varargin)
+%WAITBAR Display wait bar.
+%   H = WAITBAR(X,'message', property, value, property, value, ...)
+%   creates and displays a waitbar of fractional length X.  The
+%   handle to the waitbar figure is returned in H.
+%   X should be between 0 and 1.  Optional arguments property and
+%   value allow to set corresponding waitbar figure properties.
+%   Property can also be an action keyword 'CreateCancelBtn', in
+%   which case a cancel button will be added to the figure, and
+%   the passed value string will be executed upon clicking on the
+%   cancel button or the close figure button.
+%
+%   WAITBAR(X) will set the length of the bar in the most recently
+%   created waitbar window to the fractional length X.
+%
+%   WAITBAR(X,H) will set the length of the bar in waitbar H
+%   to the fractional length X.
+%
+%   WAITBAR(X,H,'message') will update the message text in
+%   the waitbar figure, in addition to setting the fractional
+%   length to X.
+%
+%   WAITBAR is typically used inside a FOR loop that performs a
+%   lengthy computation.
+%
+%   Example:
+%       h = waitbar(0,'Please wait...');
+%       for i=1:1000,
+%           % computation here %
+%           waitbar(i/1000,h)
+%       end
+%
+%   See also DIALOG, MSGBOX.
+
+%   Copyright 1984-2015 The MathWorks, Inc.
+
+% Error in -nojvm mode
+if ~usejava('jvm')
+    error(message('MATLAB:HandleGraphics:noJVM'));
+end
+
+inWaitbarCreate = true;
+
+if (nargin == 0)
+    error(message('MATLAB:waitbar:InvalidArguments'));
+elseif (nargin > 0)
+    % Must be a numeric value
+    if ~isnumeric(x) || ~isscalar(x)
+        error(message('MATLAB:waitbar:InvalidFirstInput'));
+    elseif ((x < 0) || (x > 1))
+        % Throw a warning in this case, clamp it down and keep going
+        % this is a behavior change and we want to eventually error out in this scenario,
+        % but want to do that gradually
+        if (x < 0)
+            x = 0;
+        elseif (x > 1)
+            x = 1;
+        end
+        % This warning will be enabled when callers no longer send in values
+        % outside the allowed range
+        % warning('MATLAB:waitbar:invalidValue', '%s\n%s%s', ...
+        %    'The first argument must be a numeric value between 0 and 1.',...
+        %    'Setting the value to: ', num2str(x));
+    end
+    
+    if (nargin == 1)
+        % A waitbar handle is not provided. Look for one.
+        f = findobj(allchild(0),'flat','Tag','TMWWaitbar');
+        
+        if isempty(f)
+            % No waitbar found. Create a new waitbar with a default string
+            name = getString(message('MATLAB:uistring:waitbar:WaitbarLabel'));
+        else
+            % Found waitbar(s). Update the first (current) waitbar with the new value
+            inWaitbarCreate = false;
+            f = f(1);
+            % Get handles from the saved appdata
+            handles = extractHandles(f);
+        end
+    elseif (nargin > 1)
+        % A waitbar message or handle to an existing waitbar has been
+        % provided
+        if ischar(whichbar) || iscellstr(whichbar)
+            % The second arg is a message, we are creating a waitbar
+            name = whichbar;
+        elseif all(ishghandle(whichbar, 'figure'))
+            % The second arg is a handle, we are updating the value
+            inWaitbarCreate = false;
+            % Get handles from the saved appdata
+            handles = extractHandles(whichbar);
+        else
[…]
