### 2017-12-14 #506 @Kevin-Mattheus-Moerman
@cMadan will you add an additional checklist for @wspr? I did this [here](https://github.com/openjournals/joss-reviews/issues/500#issuecomment-351155247) by simply manually editing the tick box comment and coping in an additional tick box list. You also need to invite the second reviewer (ping @arfon for that) so they can be assigned to the issue and tick boxes.
