### 2017-12-18 fc88974
Changed demo to run in external mode
