+++ data/temp/tempModel.txt
@@ -36,45 +36,46 @@
 
  FILES USED
 ===========================================================================
-	Input file : D:\MATLAB\GIT\GIBBON\data\temp\tempModel.feb
-	Plot file  : D:\MATLAB\GIT\GIBBON\data\temp\tempModel.xplt
-	Log file   : D:\MATLAB\GIT\GIBBON\data\temp\tempModel.txt
+	Input file : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel.feb
+	Plot file  : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel.xplt
+	Log file   : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel.txt
 
 
  MESH INFO
 ===========================================================================
 	Number of materials ............................ : 1
 	Number of domains .............................. : 1
-	Number of nodes ................................ : 10133
-	Number of solid elements ....................... : 41800
+	Number of nodes ................................ : 512
+	Number of solid elements ....................... : 343
 
 
  CONTROL DATA
 ===========================================================================
-	Module type .................................... : solid
+	Module type .................................... : biphasic
 	Analysis type .................................. : quasi-static
 	Plane strain mode .............................. : no
-	Number of timesteps ............................ : 20
-	Time step size ................................. : 0.05
+	Number of timesteps ............................ : 5100
+	Time step size ................................. : 0.001
 	Auto time stepper activated .................... : yes
 	  Optimal nr of iterations ..................... : 10
-	  Minimum allowable step size .................. : 0.0001
-	  Maximum allowable step size .................. : 0.05
+	  Minimum allowable step size .................. : 1e-05
+	  Maximum allowable step size .................. : 0.51
 	Number of loadcurves ........................... : 1
-	Number of displacement boundary conditions ..... : 3
+	Number of displacement boundary conditions ..... : 2
 	Number of concentrated nodal forces ............ : 0
-	Max nr of stiffness reformations ............... : 25
+	Max nr of stiffness reformations ............... : 15
 	per time steps
 	Max nr of Quasi-Newton iterations .............. : 0
 	between stiffness matrix reformations
 	Linesearch convergence tolerance ............... : 0.9
 	Minimum line search size ....................... : 0.01
 	Maximum number of line search iterations ....... : 5
-	Max condition number ........................... : 4.94066e-319
+	Max condition number ........................... : 0.01
 	Displacement convergence tolerance ............. : 0.001
 	Energy convergence tolerance ................... : 0.01
 	Residual convergence tolerance ................. : 0
-	Minimal residual value ......................... : 1e-020
+	Minimal residual value ......................... : 1e-20
+	Fluid pressure convergence tolerance ........... : 0.01
 
 
  OUTPUT DATA
@@ -86,26 +87,23 @@
 		Domain Variables:
 			stress               (type = mat3fs, format = ITEM)
 			relative volume      (type = float, format = ITEM)
+			effective fluid pressure (type = float, format = NODE)
+			fluid flux           (type = vec3f, format = ITEM)
 
 
  MATERIAL DATA
 ===========================================================================
-  1 - cube_mat (type: Mooney-Rivlin)
-	density ........................................... : 1
[…]
