+++ data/temp/iFEA_tempModel.txt
@@ -36,9 +36,9 @@
 
  FILES USED
 ===========================================================================
-	Input file : D:\MATLAB\GIT\GIBBON\data\temp\iFEA_tempModel.feb
-	Plot file  : D:\MATLAB\GIT\GIBBON\data\temp\iFEA_tempModel.xplt
-	Log file   : D:\MATLAB\GIT\GIBBON\data\temp\iFEA_tempModel.txt
+	Input file : /mnt/data/MATLAB/GIT/GIBBON/data/temp/iFEA_tempModel.feb
+	Plot file  : /mnt/data/MATLAB/GIT/GIBBON/data/temp/iFEA_tempModel.xplt
+	Log file   : /mnt/data/MATLAB/GIT/GIBBON/data/temp/iFEA_tempModel.txt
 
 
  MESH INFO
@@ -70,11 +70,11 @@
 	Linesearch convergence tolerance ............... : 0.9
 	Minimum line search size ....................... : 0.01
 	Maximum number of line search iterations ....... : 5
-	Max condition number ........................... : 4.94066e-319
+	Max condition number ........................... : 0.01
 	Displacement convergence tolerance ............. : 0.001
 	Energy convergence tolerance ................... : 0.01
 	Residual convergence tolerance ................. : 0
-	Minimal residual value ......................... : 1e-020
+	Minimal residual value ......................... : 1e-20
 
 
  OUTPUT DATA
@@ -137,9 +137,9 @@
 	stiffness matrix reformations = 1
 	step from line search         = 1.000000
 	convergence norms :     INITIAL         CURRENT         REQUIRED
-	   residual           2.353165e-003   1.862149e-008   0.000000e+000 
-	   energy             3.647942e-002   7.960874e-005   3.647942e-004 
-	   displacement       1.845328e+000   1.845328e+000   1.845328e-006 
+	   residual            2.353165e-03    1.862149e-08    0.000000e+00 
+	   energy              3.647942e-02    7.960874e-05    3.647942e-04 
+	   displacement        1.845328e+00    1.845328e+00    1.845328e-06 
  2
  Nonlinear solution status: time= 0.05
 	stiffness updates             = 1
@@ -147,9 +147,9 @@
 	stiffness matrix reformations = 1
 	step from line search         = 1.000000
 	convergence norms :     INITIAL         CURRENT         REQUIRED
-	   residual           2.353165e-003   3.226543e-012   0.000000e+000 
-	   energy             3.647942e-002   9.656890e-009   3.647942e-004 
-	   displacement       1.845328e+000   1.099622e-004   1.865075e-006 
+	   residual            2.353165e-03    3.226543e-12    0.000000e+00 
+	   energy              3.647942e-02    9.656890e-09    3.647942e-04 
+	   displacement        1.845328e+00    1.099622e-04    1.865075e-06 
  3
  Nonlinear solution status: time= 0.05
 	stiffness updates             = 2
@@ -157,9 +157,9 @@
 	stiffness matrix reformations = 1
 	step from line search         = 1.000000
 	convergence norms :     INITIAL         CURRENT         REQUIRED
-	   residual           2.353165e-003   1.020217e-015   0.000000e+000 
-	   energy             3.647942e-002   5.532028e-013   3.647942e-004 
-	   displacement       1.845328e+000   1.464229e-008   1.865239e-006 
+	   residual            2.353165e-03    1.020217e-15    0.000000e+00 
+	   energy              3.647942e-02    5.532028e-13    3.647942e-04 
[…]
