### 2017-12-18 gibbonCode/GIBBON#11 @siboles
testGibbon - error results in termination
@Kevin-Mattheus-Moerman openjournals/joss-reviews#506
The first encountered error when running:
`testGibbon('demo', 'test');`
Terminates the rest of the testing. Ideally, these errors would be trapped and recorded, but testing could continue.

Specifically, this occurs with:
> testGibbon -> Test file:14 of 58 DEMO_FEBio_block_biaxial_loading_stiffness_analysis.m

Yielding the trace:
>Error using  -
Matrix dimensions must agree.

>Error in patchDetach (line 12)
    FX=((FX-FX_mean)*shrinkFactor)+FX_mean;

>Error in ind2patch (line 59)
        [F,V]=patchDetach(F,V,1);

>Error in viewFourthOrderTensor (line 133)
    [Fp,Vp,Cp]=ind2patch(logicEntry,CM,'sk');

>Error in DEMO_FEBio_block_biaxial_loading_stiffness_analysis (line 286)
    viewFourthOrderTensor(c); %Visualize tensor C

>Error in run (line 96)
evalin('caller', [script ';']);

>Error in testGibbon (line 71)
            run(mFileNow);
