### 2017-12-19 6236026
Attempt to fix or at least investigate #11
