+++ data/temp/tempModel_force_out.txt
@@ -1,6 +1,6 @@
 *Title:
 *Step  = 1
-*Time  = 0.001
+*Time  = 0.1
 *Data  = Rx;Ry;Rz
 1, 0, 0, 0
 2, 0, 0, 0
@@ -182,340 +182,44 @@
 178, 0, 0, 0
 179, 0, 0, 0
 180, 0, 0, 0
-181, 0, 0, 0
-182, 0, 0, 0
-183, 0, 0, 0
-184, 0, 0, 0
-185, 0, 0, 0
-186, 0, 0, 0
-187, 0, 0, 0
-188, 0, 0, 0
-189, 0, 0, 0
-190, 0, 0, 0
-191, 0, 0, 0
-192, 0, 0, 0
-193, 0, 0, 0
-194, 0, 0, 0
-195, 0, 0, 0
-196, 0, 0, 0
-197, 0, 0, 0
-198, 0, 0, 0
-199, 0, 0, 0
-200, 0, 0, 0
-201, 0, 0, 0
-202, 0, 0, 0
-203, 0, 0, 0
-204, 0, 0, 0
-205, 0, 0, 0
-206, 0, 0, 0
-207, 0, 0, 0
-208, 0, 0, 0
-209, 0, 0, 0
-210, 0, 0, 0
-211, 0, 0, 0
-212, 0, 0, 0
-213, 0, 0, 0
-214, 0, 0, 0
-215, 0, 0, 0
-216, 0, 0, 0
-217, 0, 0, 0
-218, 0, 0, 0
-219, 0, 0, 0
-220, 0, 0, 0
-221, 0, 0, 0
-222, 0, 0, 0
-223, 0, 0, 0
-224, 0, 0, 0
-225, 0, 0, 0
-226, 0, 0, 0
-227, 0, 0, 0
-228, 0, 0, 0
-229, 0, 0, 0
-230, 0, 0, 0
-231, 0, 0, 0
-232, 0, 0, 0
-233, 0, 0, 0
-234, 0, 0, 0
-235, 0, 0, 0
-236, 0, 0, 0
-237, 0, 0, 0
-238, 0, 0, 0
-239, 0, 0, 0
-240, 0, 0, 0
-241, 0, 0, 0
-242, 0, 0, 0
-243, 0, 0, 0
-244, 0, 0, 0
-245, 0, 0, 0
-246, 0, 0, 0
-247, 0, 0, 0
-248, 0, 0, 0
-249, 0, 0, 0
-250, 0, 0, 0
-251, 0, 0, 0
-252, 0, 0, 0
-253, 0, 0, 0
-254, 0, 0, 0
-255, 0, 0, 0
-256, 0, 0, 0
-257, 0, 0, 0
-258, 0, 0, 0
-259, 0, 0, 0
-260, 0, 0, 0
-261, 0, 0, 0
-262, 0, 0, 0
-263, 0, 0, 0
-264, 0, 0, 0
-265, 0, 0, 0
-266, 0, 0, 0
-267, 0, 0, 0
-268, 0, 0, 0
-269, 0, 0, 0
-270, 0, 0, 0
-271, 0, 0, 0
-272, 0, 0, 0
-273, 0, 0, 0
-274, 0, 0, 0
-275, 0, 0, 0
-276, 0, 0, 0
-277, 0, 0, 0
-278, 0, 0, 0
-279, 0, 0, 0
-280, 0, 0, 0
-281, 0, 0, 0
-282, 0, 0, 0
-283, 0, 0, 0
-284, 0, 0, 0
-285, 0, 0, 0
-286, 0, 0, 0
-287, 0, 0, 0
-288, 0, 0, 0
[…]
