+++ data/temp/tempModel.txt
@@ -45,23 +45,23 @@
 ===========================================================================
 	Number of materials ............................ : 1
 	Number of domains .............................. : 1
-	Number of nodes ................................ : 512
-	Number of solid elements ....................... : 343
+	Number of nodes ................................ : 216
+	Number of solid elements ....................... : 125
 
 
  CONTROL DATA
 ===========================================================================
-	Module type .................................... : biphasic
+	Module type .................................... : solid
 	Analysis type .................................. : quasi-static
 	Plane strain mode .............................. : no
-	Number of timesteps ............................ : 5100
-	Time step size ................................. : 0.001
+	Number of timesteps ............................ : 10
+	Time step size ................................. : 0.1
 	Auto time stepper activated .................... : yes
 	  Optimal nr of iterations ..................... : 10
 	  Minimum allowable step size .................. : 1e-05
-	  Maximum allowable step size .................. : 0.51
+	  Maximum allowable step size .................. : 0.1
 	Number of loadcurves ........................... : 1
-	Number of displacement boundary conditions ..... : 2
+	Number of displacement boundary conditions ..... : 1
 	Number of concentrated nodal forces ............ : 0
 	Max nr of stiffness reformations ............... : 15
 	per time steps
@@ -75,7 +75,6 @@
 	Energy convergence tolerance ................... : 0.01
 	Residual convergence tolerance ................. : 0
 	Minimal residual value ......................... : 1e-20
-	Fluid pressure convergence tolerance ........... : 0.01
 
 
  OUTPUT DATA
@@ -87,23 +86,25 @@
 		Domain Variables:
 			stress               (type = mat3fs, format = ITEM)
 			relative volume      (type = float, format = ITEM)
-			effective fluid pressure (type = float, format = NODE)
-			fluid flux           (type = vec3f, format = ITEM)
+			shell thickness      (type = float, format = COMP)
 
 
  MATERIAL DATA
 ===========================================================================
-  1 - Block_material (type: biphasic)
-	phi0 .............................................. : 0.5
-	fluid_density ..................................... : 1e-09
+  1 - mat_1 (type: uncoupled solid mixture)
+	density ........................................... : 1
+	k ................................................. : 101
+	laugon ............................................ : 0
+	atol .............................................. : 0.01
+	minaug ............................................ : 0
+	maxaug ............................................ : 0
 
 
  LOADCURVE DATA
 ===========================================================================
   1
          0         0
-       0.1         1
-       5.1         1
+         1         1
 
 
  LINEAR SOLVER DATA
@@ -113,1072 +114,1576 @@
 
 
 
-===== beginning time step 1 : 0.001 =====
+===== beginning time step 1 : 0.1 =====
 ===== reforming stiffness matrix:
-	Nr of equations ........................... : 1968
-	Nr of nonzeroes in stiffness matrix ....... : 81948
+	Nr of equations ........................... : 600
+	Nr of nonzeroes in stiffness matrix ....... : 17024
 
  1
- Nonlinear solution status: time= 0.001
[…]
