### 2017-12-19 gibbonCode/GIBBON#11 @Kevin-Mattheus-Moerman
I cannot reproduce this error. What FEBio version do you use?

Can you run the demo on its own? (i.e. without using testGibbon)
`DEMO_FEBio_block_biaxial_loading_stiffness_analysis`

Can you check if this file is being created and if you have delete/write permission in that region, you can type this in the command window to see where the file is:
`fullfile(savePath,FEB_struct.run_output_names{3})`
That file should be the FEBio log file containing the stiffness tensor entries. I wonder if it is empty or if it does not contain the right amount of entries for some reason.

Just before the error (line 285 ish) could you check the size of `c` by typing `size(c)`

If you update GIBBON you'll see that I've made two changes to check something.
1) I switched to _external_ running mode (line 241):
`FEBioRunStruct.runMode='external';%'internal';`
This is just to check if for some reason Arch linux does not like the "internal" mode. I noticed this demo was using internal mode.

2) I increased the maximum log file checking time (line 244):
`FEBioRunStruct.maxLogCheckTime=10; %Max log file checking time `
This should not be the error but just in case.

> Terminates the rest of the testing. Ideally, these errors would be trapped and recorded, but testing could continue.

I agree that would be nice. I'll consider that/think about how to do that. For now the user can read the message displayed which says "Test file: 14 of 58" failed. The user can then start from this file by setting startInd (start index) to 14.

`startInd=14; %The start index`
`approveOpt=0; %Wait to approve every time, yes=1, no=0`
`testGibbon('demo','test',approveOpt,startInd);`
