### 2017-12-19 gibbonCode/GIBBON#11 @siboles
I am using FEBio 2.5.2.8980. I pulled your changes and reinstalled (not sure if that is necessary or not.) The error still occurs.

The stiffness tensor was being saved in ../GIBBON/data/temp/tempModel_stiffness_out.txt.

The file is not empty, and size(c) was 3x3x3x3.
I had a colleague install and run this on a Windows machine and the test passed. He is currently running the rest of the DEMOs. I will post the results of those as a separate issue if there are any errors.
