### 2017-12-19 gibbonCode/GIBBON#12 @siboles
Windows - Record of all DEMO_ runs
@Kevin-Mattheus-Moerman openjournals/joss-reviews#506

My colleague, @azimj, ran all the demos on a Windows machine. He has logged the command window output in the attached file  - [run_tests_201712191554.txt](https://github.com/gibbonCode/GIBBON/files/1573648/run_tests_201712191554.txt).

Seems all passed but two.
DEMO_FEBio_beam_L_force.m (this passed for me)

and

DEMO_FEBio_skin_sphere_indentation_tet4_vs_tet10.m (FEBio Error Termination)
