### 2017-12-20 gibbonCode/GIBBON#11 @siboles
My issues all seem to stem from this ind2patch function. Had the same problem with meshing demos.

>Error using  -
Matrix dimensions must agree.

>Error in patchDetach (line 12)
    FX=((FX-FX_mean)*shrinkFactor)+FX_mean;

>Error in ind2patch (line 59)
        [F,V]=patchDetach(F,V,1);

>Error in DEMO_MixedTetHexMeshing_torus (line 70)
[Fm,Vm,Cm]=ind2patch(L_plot,double(M),'sk');

>Error in run (line 96)
evalin('caller', [script ';']);

>Error in testGibbon (line 71)
            run(mFileNow);
