### 2017-12-20 097d889
Added control settings to select demos to address #12
DEMO_FEBio_beam_L_force and DEMO_FEBio_skin_sphere_indentation_tet4_vs_tet10 now have control parameters at the top so users can adjust them to improve model stability. Specifically using full-newton iterations may help.
