+++ docs/DEMO_FEBio_beam_L_force.m
@@ -53,6 +53,15 @@ v2=0.3;
 
 nRefine=0;
 
+%FEA control settings
+nSteps=20; %Number of time steps desired
+max_refs=25; %Max reforms
+max_ups=0; %Set to zero to use full-Newton iterations
+opt_iter=10; %Optimum number of iterations
+max_retries=5; %Maximum number of retires
+dtmin=(1/nSteps)/100; %Minimum time step size
+dtmax=1/nSteps; %Maximum time step size
+
 %% CREATING MESHED BOX
 
 %Create box 1
@@ -281,14 +290,16 @@ FEB_struct.Control.Properties={'time_steps','step_size',...
     'max_refs','max_ups',...
     'dtol','etol','rtol','lstol'};
 
-numAnalysisSteps=20; 
-timeStepSize=1/numAnalysisSteps;
-
-FEB_struct.Control.Values={numAnalysisSteps,timeStepSize,...
-    25,10,...
+%Control section
+FEB_struct.Control.AnalysisType='static';
+FEB_struct.Control.Properties={'time_steps','step_size',...
+    'max_refs','max_ups',...
+    'dtol','etol','rtol','lstol'};
+FEB_struct.Control.Values={nSteps,1/nSteps,...
+    max_refs,max_ups,...
     0.001,0.01,0,0.9};
 FEB_struct.Control.TimeStepperProperties={'dtmin','dtmax','max_retries','opt_iter'};
-FEB_struct.Control.TimeStepperValues={timeStepSize/100,timeStepSize, 5, 10};
+FEB_struct.Control.TimeStepperValues={dtmin,dtmax,max_retries,opt_iter};
 
 %Defining node sets
 FEB_struct.Geometry.NodeSet{1}.Set=bcSupportList;
