+++ data/temp/tempModel.txt
@@ -15,7 +15,7 @@
                                                                            
       F I N I T E   E L E M E N T S   F O R   B I O M E C H A N I C S      
                                                                            
-                 --- v e r s i o n - 2 . 5 . 2 . 8980 ---                 
+                 --- v e r s i o n - 2 . 6 . 4 . 9579 ---                 
                                                                            
                                                                            
   Musculoskeletal Research Laboratory                                      
@@ -23,15 +23,20 @@
   http://febio.org                                                         
                                                                            
   FEBio is a registered trademark.                                         
-  copyright (c) 2006-2015 - All rights reserved                            
+  copyright (c) 2006-2016 - All rights reserved                            
                                                                            
- This is the NON-COMMERCIAL version of FEBio. This version may only be          
- used for non-commercial purposes as described in the license agreement.        
- To obtain a valid commercial license file, please contact the developers.      
+ This is the NON-COMMERCIAL version of FEBio. This version may only be  
+ used for non-commercial purposes as described in the license agreement.
+ To obtain a valid commercial license file, please contact the developers.
                                                                            
 ===========================================================================
 
 
+ *************************************************************************
+ *                               WARNING                                 *
+ *                                                                       *
+ * 3200 isolated vertices removed.                                       *
+ *************************************************************************
 
 
  FILES USED
@@ -43,10 +48,10 @@
 
  MESH INFO
 ===========================================================================
-	Number of materials ............................ : 1
-	Number of domains .............................. : 1
-	Number of nodes ................................ : 216
-	Number of solid elements ....................... : 125
+	Number of materials ............................ : 2
+	Number of domains .............................. : 2
+	Number of nodes ................................ : 3720
+	Number of solid elements ....................... : 225
 
 
  CONTROL DATA
@@ -54,16 +59,16 @@
 	Module type .................................... : solid
 	Analysis type .................................. : quasi-static
 	Plane strain mode .............................. : no
-	Number of timesteps ............................ : 10
-	Time step size ................................. : 0.1
+	Number of timesteps ............................ : 20
+	Time step size ................................. : 0.05
 	Auto time stepper activated .................... : yes
 	  Optimal nr of iterations ..................... : 10
-	  Minimum allowable step size .................. : 1e-05
-	  Maximum allowable step size .................. : 0.1
+	  Minimum allowable step size .................. : 0.0005
+	  Maximum allowable step size .................. : 0.05
 	Number of loadcurves ........................... : 1
-	Number of displacement boundary conditions ..... : 1
-	Number of concentrated nodal forces ............ : 0
-	Max nr of stiffness reformations ............... : 15
[…]
