+++ config/FEBioPath.txt
@@ -1,2 +1,2 @@
 #List paths for febio here, multiple can be given e.g. for use on several platforms. The first valid path found is always used.
-/home/kevin/febio-2.5.2/bin/febio2.lnx64
+/home/kevin/febio-2.6.4/bin/febio2.lnx64
