### 2017-12-20 gibbonCode/GIBBON#12 @Kevin-Mattheus-Moerman
Thanks for catching these. I think these are simply FEBio issues. The first model is force controlled and the second involves contact and non-linear elements. Both could benefit from altered FEA control parameters and it may be that they solve in one FEBio version and not in another. For instance I think using full-Newton iterations will help with these. To use that set max_ups to 0. See below. I've been running these on the latest FEBio version (2.6.4).

For both demos I've now added a set of FEA control parameters at the top so the user can alter them more easily. There may be other e.g. contact related demos that could be "fragile" in this sense so I'll likely update those with similar settings if this seems to solve it.

`%FEA control settings`
`nSteps=20; %Number of time steps desired`
`max_refs=25; %Max reforms`
`max_ups=0; %Set to zero to use full-Newton iterations`
`opt_iter=10; %Optimum number of iterations`
`max_retries=5; %Maximum number of retires`
`dtmin=(1/nSteps)/100; %Minimum time step size`
`dtmax=1/nSteps; %Maximum time step size`
