+++ docs/DEMO_FEBio_disc_probe_indentation.m
@@ -277,10 +277,7 @@ inputStruct.sizeData=edgeSizeField;
 % <http://wias-berlin.de/software/tetgen/>)
 
 [meshOutput]=runTetGen(inputStruct); %Run tetGen
-
-%%
-% Call TetView to view the model
-% runTetView(meshOutput.loadNameStruct.loadName_ele);
+meshView(meshOutput);
 
 %%
 % Access model element and patch data
