+++ docs/DEMO_vessel_FEBIO_fibre_mapping.m
@@ -1,616 +1,0 @@
-%% DEMO_vessel_FEBIO_fibre_mapping
-% Below is a demonstration of how to map, set and import fibre directions
-% for FEBio.
-
-%%
-% A vessel model is constructed consisting of two layers with different
-% mechanical properties and fibre directions. The inner wall of the vessel
-% is subjected to a prescribed displacement. Fibre directions are set in
-% the .feb file, the model is run in FEBio, results are exported and the
-% fibre directions in the deformed state are derived from the imported
-% model results. 
-
-%%
-
-clear; close all; clc;
-
-%%
-% Plot settings
-fontSize=15;
-faceAlpha1=0.5;
-faceAlpha2=0.5;
-edgeColor=0.25*ones(1,3);
-edgeWidth=1.5;
-markerSize1=50; 
-cMap=gjet(250);
-
-% path names
-defaultFolder = fileparts(fileparts(mfilename('fullpath')));
-savePath=fullfile(defaultFolder,'data','temp');
-
-%%
-
-%Material parameters
-c1_1=0.0006618290645;
-m1_1=2.0000011857416;
-ksi_1=0.0000906045135;
-beta_1=3.1728247084040;
-f_1=229.9578182262337;
-k_factor=500; 
-k_1=(c1_1+ksi_1)*k_factor;
-
-c1_2=2*0.0006618290645;
-m1_2=2.0000011857416;
-ksi_2=2*0.0000906045135;
-beta_2=3.1728247084040;
-f_2=229.9578182262337;
-k_factor=500; 
-k_2=(c1_2+ksi_2)*k_factor;
-
-%% BUILDING EXAMPLE SURFACE GEOMETRY
-
-% Construct regular cylinder
-r=10; 
-nr=25; 
-h=35;
-nz=15;
-ptype='tri';
-[Fc,Vc]=patchcylinder(r,nr,h,nz,ptype);
-
-%Top edge indices
-It=ones(nr,1); Jt=(1:1:nr)';
-INDt=sub2ind([nz,nr],It,Jt);
-
-%Bottom edge indices
-Ib=nz.*ones(nr,1); Jb=(1:1:nr)';
-INDb=sub2ind([nz,nr],Ib,Jb);
-
-% Deforming regular cylinder to create outer wall
-V1=Vc; 
-F1=Fc; 
-
-% Creating the first inner wall
-[T,R,Z]=cart2pol(V1(:,1),V1(:,2),V1(:,3));
-R2=R-2;
-V2=V1;
-F2=Fc; 
-[V2(:,1),V2(:,2),V2(:,3)]=pol2cart(T,R2,Z);
-
-% Creating the second inner wall
-[T,R,Z]=cart2pol(V2(:,1),V2(:,2),V2(:,3));
-R3=R-2;
-V3=V2;
-F3=Fc; 
-[V3(:,1),V3(:,2),V3(:,3)]=pol2cart(T,R3,Z);
-
-%%
-% Plotting the example model surfaces
-hf1=cFigure;
-title('Cylindrical surfaces of the model','FontSize',fontSize);
-xlabel('X','FontSize',fontSize);ylabel('Y','FontSize',fontSize); zlabel('Z','FontSize',fontSize);
-hold on;
-patch('Faces',F1,'Vertices',V1,'EdgeColor','k','FaceColor','r','FaceAlpha',1);
-patch('Faces',F2,'Vertices',V2,'EdgeColor','k','FaceColor','g','FaceAlpha',1);
-patch('Faces',F3,'Vertices',V3,'EdgeColor','k','FaceColor','b','FaceAlpha',1);
-
-% patchNormPlot(F1,V1);
-% patchNormPlot(F2,V2);
-% patchNormPlot(F3,V3);
-
[…]
