+++ docs/DEMO_FEBio_cylinder_embedded_indentor_01.m
@@ -194,10 +194,7 @@ inputStruct.modelName=modelName;
 % <http://wias-berlin.de/software/tetgen/>)
 
 [meshOutput]=runTetGen(inputStruct); %Run tetGen 
-
-%%
-% Call TetView to view the model
-runTetView(meshOutput.loadNameStruct.loadName_ele);
+meshView(meshOutput);
 
 %% 
 % Access model element and patch data 

[…]
