### 2017-12-20 gibbonCode/GIBBON#12 @siboles
I would recommend running all the demos in full-Newton. The consistency would be worth the trade-off in efficiency. For contact, I always found facet-to-facet sliding, two pass, using penalty rather than Langrange multipliers to be more robust. Of course the penalty method will exhibit some penetration depending on how high relative to the material properties one sets the value. Likewise, for the biphasic contact sliding2 constraint. I think it is fair to close this issue now though.
