### 2017-12-20 gibbonCode/GIBBON#11 @siboles
I don't think this issue should impede the acceptance for JOSS, since it is OS specific. I will close this, and recommend opening a new issue without linking to the review.
