### 2017-12-21 60d406c
Targets #14 and #15
patchMarchDistMap now uses experimental code if mex file code throws error. Also all FEBio demos now use full-Newton iterations and allow users to alter this based on control parameters.
