+++ docs/DEMO_FEBio_beam_L_force.m
@@ -53,14 +53,14 @@ v2=0.3;
 
 nRefine=0;
 
-%FEA control settings
-nSteps=20; %Number of time steps desired
+% FEA control settings
+numTimeSteps=20; %Number of time steps desired
 max_refs=25; %Max reforms
 max_ups=0; %Set to zero to use full-Newton iterations
 opt_iter=10; %Optimum number of iterations
 max_retries=5; %Maximum number of retires
-dtmin=(1/nSteps)/100; %Minimum time step size
-dtmax=1/nSteps; %Maximum time step size
+dtmin=(1/numTimeSteps)/100; %Minimum time step size
+dtmax=1/numTimeSteps; %Maximum time step size
 
 %% CREATING MESHED BOX
 
@@ -295,7 +295,7 @@ FEB_struct.Control.AnalysisType='static';
 FEB_struct.Control.Properties={'time_steps','step_size',...
     'max_refs','max_ups',...
     'dtol','etol','rtol','lstol'};
-FEB_struct.Control.Values={nSteps,1/nSteps,...
+FEB_struct.Control.Values={numTimeSteps,1/numTimeSteps,...
     max_refs,max_ups,...
     0.001,0.01,0,0.9};
 FEB_struct.Control.TimeStepperProperties={'dtmin','dtmax','max_retries','opt_iter'};
