+++ docs/DEMO_FEBio_strip_multi_step_clamp.m
@@ -65,6 +65,15 @@ f_ini=235.1;
 k_factor=100;
 k_ini=(2*c1_ini+ksi_ini)*k_factor;
 
+% FEA control settings
+numTimeSteps=10; %Number of time steps desired
+max_refs=25; %Max reforms
+max_ups=0; %Set to zero to use full-Newton iterations
+opt_iter=6; %Optimum number of iterations
+max_retries=5; %Maximum number of retires
+dtmin=(1/numTimeSteps)/100; %Minimum time step size
+dtmax=1/numTimeSteps; %Maximum time step size
+
 %% CREATING 3 MESHED BOXES
 
 %Create box 1
@@ -222,8 +231,6 @@ drawnow;
 
 %% CONSTRUCTING FEB MODEL
 
-%% CONSTRUCTING FEB MODEL
-
 FEB_struct.febio_spec.version='2.0';
 
 % Defining file names
@@ -282,11 +289,12 @@ FEB_struct.Step{1}.Control.AnalysisType='static';
 FEB_struct.Step{1}.Control.Properties={'time_steps','step_size',...
     'max_refs','max_ups',...
     'dtol','etol','rtol','lstol'};
-FEB_struct.Step{1}.Control.Values={20,0.05,...
-    25,0,...
+FEB_struct.Step{1}.Control.Values={numTimeSteps,1/numTimeSteps,...
+    max_refs,max_ups,...
     0.001,0.01,0,0.9};
 FEB_struct.Step{1}.Control.TimeStepperProperties={'dtmin','dtmax','max_retries','opt_iter'};
-FEB_struct.Step{1}.Control.TimeStepperValues={1e-4,0.05,5,10};
+FEB_struct.Step{1}.Control.TimeStepperValues={dtmin,dtmax,max_retries,opt_iter};
+
 FEB_struct.Step{2}.Control=FEB_struct.Step{1}.Control;
 
 %Step specific BC's
