### 2017-12-21 gibbonCode/GIBBON#12 @Kevin-Mattheus-Moerman
All demos now run in full-Newton mode
