### 2017-12-21 gibbonCode/GIBBON#11 @Kevin-Mattheus-Moerman
Handling of errors in testGibbon has been improved. It now catches the error and gives you the file index so you can resume from that point
