### 2017-12-21 gibbonCode/GIBBON#11 @Kevin-Mattheus-Moerman
@siboles have you tried `testGibbon('help', 'test');`? Do all of those run?
Have you tried `HELP_im2patch` and `HELP_ind2patch`. I suspect that `size(FX_mean)` does not match `size(FX)` for you in `patchDetach`. Can you report on those sizes? I wonder if you need a transpose somewhere...
