### 2017-12-22 #506 @Kevin-Mattheus-Moerman
@wspr let me know if I can help with anything. See `testGibbon` to run all tests.
