### 2017-12-22 #506 @cMadan
@siboles, is it safe to assume that you are now satisfied with the submission and endorse it's acceptance in the current form?

@Kevin-Mattheus-Moerman, I'll still wait for @wspr, but thought I'd get that confirmation from @siboles, rather than checking later on.
