### 2017-12-23 #506 @wspr
Hi all — the end of the year hit me rather suddenly so I haven’t had a chance to fully investigate. If you could wait a little longer (say, before the end of the year) I’d be very happy to put in my 2c as well, but if you’d rather knock it on the head immediately I am happy to give it a miss. Sorry for the hold up.
