### 2017-12-23 #506 @cMadan
@wspr, I have no problem waiting a bit longer. Also feel free to take some time into January too (I'd hate to have reviewing take away from holiday plans).

That said, can you check if that second checklist works for you? If not, I'll have to see about permissions...
