### 2017-12-24 #506 @arfon
@wspr - if you accept the invite link here: https://github.com/openjournals/joss-reviews/invitations you should be able to edit the checklist.
