### 2017-12-26 #506 @siboles
Very sorry folks for the slow reply. Yes I am satisfied with GIBBON and
vote to accept it for publication.
