### 2018-01-12 9bff5b5
Removed polyrevolve (replaced by polyRevolve) and fixed minor documentation rendering issues on website
