+++ lib/polyrevolve.m
@@ -1,112 +1,0 @@
-function [X,Y,Z]=polyrevolve(x,z,n)
-
-% function [X,Y,Z]=polyrevolve(x,z,n)
-% ------------------------------------------------------------------------
-% This function revolves a 2D polygon around the Z-axis.
-%
-% x is a vector with x coordinates
-% z is a vector with z coordinates
-% n is the resolution of the rotation in radians
-%
-% The vectors x and z are converted to spherical coordinates (rho,r). The
-% points are then 'copied' around the z-axis. The space between each point
-% is defined by n.
-%
-% %EXAMPLE
-% clear all; close all; clc;
-% x=0:0.2:2*pi;
-% z=sin(x+(pi/2));
-% revolve_res=0.2;
-% [X,Y,Z]=polyrevolve(x,z,revolve_res);
-%
-% grid_res=0.05;
-% [XI,YI] = meshgrid(min(min(X)):grid_res:max(max(X)),min(min(Y)):grid_res:max(max(Y)));
-% ZI = griddata(X,Y,Z,XI,YI,'cubic');
-% figure; fig=gcf; clf(fig); units=get(fig,'units'); set(fig,'units','normalized','outerposition',[0 0 1 1]); set(fig,'units',units);
-% plot(x,z,'r-'); hold on; axis equal;
-% xlabel('x (mm)'); ylabel('z (mm)');
-% title('The polygon');
-% xlabel('x (mm)'); ylabel('z (mm)');
-% figure; fig=gcf; clf(fig); units=get(fig,'units'); set(fig,'units','normalized','outerposition',[0 0 1 1]); set(fig,'units',units);
-% plot3(X,Y,Z,'k.'); hold on; axis equal;
-% title('Data "copied" around z-axis using POLYREVOLVE');
-% xlabel('x (mm)'); ylabel('y (mm)'); zlabel('z (mm)');
-% figure; fig=gcf; clf(fig); units=get(fig,'units'); set(fig,'units','normalized','outerposition',[0 0 1 1]); set(fig,'units',units);
-% surf(XI,YI,ZI,'EdgeColor','none'); hold on; axis equal;
-% shading interp; material shiny; lightangle(45,30);
-% title('Surface fit of the revolved data points')
-% xlabel('x (mm)'); ylabel('y (mm)'); zlabel('z (mm)');
-%
-%%EXAMPLE 2
-% clear all; close all; clc;
-% x=[40 65];
-% z=0;
-% no_markers=6;
-% revolve_res=2*x(1)*sin(pi/no_markers);
-% [X,Y,Z]=polyrevolve(x,z,revolve_res);
-% Xr=round(X);
-% Yr=round(Y);
-% Zr=round(Z);
-% figure; fig=gcf; clf(fig); units=get(fig,'units'); set(fig,'units','normalized','outerposition',[0 0 1 1]); set(fig,'units',units);
-% h=plot(X,Y,'r.'); hold on; axis equal; axis([min(X)-1 max(X)+1 min(Y)-1 max(Y)+1]);
-% set(h,'MarkerSize',20);
-% h=plot(Xr,Yr,'b.'); hold on; axis equal; axis([min(X)-1 max(X)+1 min(Y)-1 max(Y)+1]);
-% set(h,'MarkerSize',20);
-%
-% Kevin Mattheus Moerman
-% kevinmoerman@hotmail.com
-% 14/08/2008
-% ------------------------------------------------------------------------
-
-[theta, rho, r]=cart2sph(x,zeros(size(x)),z);
-THETA=[];
-RHO=[];
-R=[];
-for i= 1: length(r)
-    if n>(2*x(i))
-        %The distance between points on a circle can not exceed the
[…]
