### 2018-01-15 gibbonCode/GIBBON#17 @wspr
Installation: best way to chmod +x where necessary?
I tried running the example file `/DEMO_MixedTetHexMeshing.m` and at one point it gracefully failed with:

```
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
--- Writing SMESH file --- 15-Jan-2018 14:03:11
----> Adding node field
----> Adding facet field
----> Adding holes specification
----> Adding region specification
--- Done --- 15-Jan-2018 14:03:11
--- Running TetGen to mesh input boundary--- 15-Jan-2018 14:03:11
/bin/bash: /Users/will/Repo/External/GIBBON/lib_ext/tetGen/mac64/tetgen: Permission denied
--- Done --- 15-Jan-2018 14:03:11
```

Now, that's an easy fix in my case:

    chmod +x /Users/will/Repo/External/GIBBON/lib_ext/tetGen/mac64/tetgen

But I wonder whether it would make sense to somehow automate this for all needed files. Are there best practices for doing this within Matlab itself? (I'm guessing the chmod settings aren't reliably preserved after a git clone but I could be wrong.)
