### 2018-01-15 gibbonCode/GIBBON#17 @wspr
Actually it looks like files in git should preserve their chmod-ness, and the following should work (famous last words) across platforms:

    git update-index --chmod=+x path/to/tetgen
