### 2018-01-16 gibbonCode/GIBBON#17 @Kevin-Mattheus-Moerman
@wspr thanks for your work on this.  What do you think is the best way to address this? I've had users where MATLAB and/or FEBio and/or TetGen presented with file permission issues (since all create and modify files and also GIBBON cleans these up by deleting temporary files). Am I correct in saying you only had file permission issues with TetGen itself? So the fix you present addresses that? Can you advise as to how I should address this issue, or say if my attempt described below is sufficient?

I've now added a statement in the [README](https://github.com/gibbonCode/GIBBON/blob/master/README.md) to attempt to address potential file permission issues (see the first step in the installation process described):

---
### 1. Get a copy of GIBBON
To use GIBBON you need to create a copy on a local directory on your machine. Obtain a copy by downloading and unzipping the latest [zip file](https://github.com/gibbonCode/GIBBON/archive/master.zip) or clone GIBBON instead e.g. using: `git clone https://github.com/gibbonCode/GIBBON.git`. You can place the GIBBON folder anywhere on your machine, however, MATLAB (and the 3rd party packages listed below) may have file permission limitations for some locations which you may need to address<sup>\*</sup>.

<sup>\* For instance, some of GIBBON's features, such as those associated with TetGen and FEBio, regularly create and delete temporary files. As such MATLAB (and the 3rd party packes) should have full permissions for these folders. First of all users should make sure MATLAB has full (e.g. read/write/delete) permissions for the entire GIBBON folder (and its sub-folders). On some platforms the 3rd party packages require special treatment. For instance some OSX users have found it necessary to give tetGen or FEBio particular file permissions. For instance in the case of TetGen related features using `git update-index --chmod=+x path/to/tetgen`. </sup>
