### 2018-01-16 462b826
Addresses  #17. Updated README with advice on file permission issues
