### 2018-01-16 5fa2783
Ran testGibbon to update documentation for all HELP_ files
