### 2018-01-17 gibbonCode/GIBBON#17 @wspr
Thanks for the detailed reply, the text you've written looks good.

Sorry I was unclear above, though — my understanding (although it's hard for me to test this) is that if you write

    git update-index --chmod=+x path/to/your/tetgen

and push the resulting commit to Github it will then solve the problem for everyone. I tried looking (non-exhaustively, I want to emphasise!) to see if there were any other files that needed this done and came up blank, so I might have been just (un)lucky to stumble upon this early in testing.

If you give that a try I should be able to confirm.
