### 2018-01-19 c1cdbb2
Targets #19, create febioStruct2xml for general XML file support, full FEBio compatibility, and compliance with febio_spec 2.5
