+++ docs/DEMO_FEBio_bar_force.m
@@ -50,6 +50,10 @@ max_retries=5; %Maximum number of retires
 dtmin=(1/nSteps)/100; %Minimum time step size
 dtmax=1/nSteps; %Maximum time step size
 
+%Material parameters
+c1=1e-3;
+k=1e3*c1;
+
 %% CREATING MESHED BOX
 
 %Create box 1
@@ -120,8 +124,6 @@ FEB_struct.Geometry.ElementMat={elementMaterialIndices};
 FEB_struct.Geometry.ElementsPartName={'Bar'};
 
 %Material section
-c1=1e-3;
-k=1e3*c1;
 FEB_struct.Materials{1}.Type='Mooney-Rivlin';
 FEB_struct.Materials{1}.Properties={'c1','c2','k'};
 FEB_struct.Materials{1}.Values={c1,0,k};
