### 2018-01-21 #506 @Kevin-Mattheus-Moerman
I've been working on @wspr 's comments. He had opened some issues on my repo which I believe I've addressed. There may be more to come but I'll let him comment on that.
