### 2018-01-21 #506 @wspr
@cMadan — sorry this dragged on a little; I think I'm basically there.

@Kevin-Mattheus-Moerman — did you see my last message for testing out the `git update-index --chmod=+x lib_ext/tetGen/mac64/tetgen` thing? I believe if you run that on your end then commit, I'll be able to check and confirm. The one thing I'm not sure about is how many more of these are needed; I guess it can't hurt to also run on the `lin64` script as well.
