### 2018-01-22 gibbonCode/GIBBON#20 @wspr
Paper review
https://github.com/openjournals/joss-reviews/issues/506

I'm really impressed by the breadth and ease-of-installation of this comprehensive Matlab toolbox. Thanks for the great effort!

For the text of the paper, I have minor tidbits to suggestion improvements; let me know if anything is unclear:

* Possible repetition: (but okay if you want to keep it for emphasis)

    > an **open-source** MATLAB® toolbox providing a single **open-source** framework

* Missing comma:

    > ‘Using GIBBON geometry’ → ‘Using GIBBON, geometry’

* Missing space:

    > continuum mechanics[@Moerman2017a]

* Perhaps word improvement:

    > ‘However, several CAD style commands have also been implemented in GIBBON’
 → ‘In addition, for generating simple geometries within MATLAB,  GIBBON also provides several CAD-style commands’

* Typo: ‘smoothening’

* Can be simplified:

    > ‘Lattice structures: Various lattice structure tools have been implemented.’ → could be just, e.g., ‘Lattice structure tools:’. Or if the opening sentence is kept, use it to explain what contexts the lattice structure tools are used in.

* FEBio:

  * `GIBBON interfaces with the free software [FEBio](http://febio.org/) [@Maas2012] for FEA` — should probably spell out FEA here.

  * `GIBBON can be used as a pre- and post- processor for FEA` — or for FEBio?

  * `it enables the code based development` → ‘ it enables code-based development’

* `Visualization` — this dot point is inconsistent with the rest; should have a colon and no newline.

* Nested references end up having double parentheses; not sure if I'd prefer dropping the outer ones? I understand this probably isn't quite as flexible as a standard LaTeX workflow `:)`

    > ([@Cooney2015] [@Takaza2013])

* ‘National Institute of Health (R01EB024531-01) presently’ — start with "and" as it's the last in the list.
