### 2018-01-22 #506 @cMadan
Thank you for the review, @wspr! I think you're all done.

@Kevin-Mattheus-Moerman, let me know when you've made these minor changes to the paper. After that's done, please let me know the DOI for the archive.
