### 2018-01-24 5d5a3bb
Updated febioStruct2xml and created associated documentation
