+++ docs/DEMO_febio_0001_cube_uniaxial.m
@@ -1,0 +1,364 @@
+%% DEMO_febio_0001_cube_uniaxial
+% Below is a demonstration for:
+% 1) Building an FEBio model for uniaxial loading
+% 2) Running the model
+% 3) Importing displacement and force results
+
+%%
+
+clear; close all; clc;
+
+%%
+% Plot settings
+fontSize=20;
+faceAlpha1=0.8;
+faceAlpha2=1;
+edgeColor=0.25*ones(1,3);
+edgeWidth=1.5;
+markerSize=40;
+lineWidth=3;
+
+%%
+% Control parameters
+
+% path names
+defaultFolder = fileparts(fileparts(mfilename('fullpath')));
+savePath=fullfile(defaultFolder,'data','temp');
+
+% Defining file names
+modelNameEnd='tempModel';
+modelName=fullfile(savePath,modelNameEnd);
+
+febioFebFileName=[modelName,'.feb']; %FEB file name
+febioLogFileName=[modelName,'.txt']; %FEBio log file name
+febioLogFileName_disp=[modelNameEnd,'_disp_out.txt']; %Log file name for exporting force
+febioLogFileName_force=[modelNameEnd,'_force_out.txt']; %Log file name for exporting force
+febioLogFileName_stress=[modelNameEnd,'_stress_out.txt']; %Log file name for exporting force
+
+%Specifying dimensions and number of elements
+sampleWidth=10;
+sampleThickness=10; 
+sampleHeight=10;
+pointSpacings=2*ones(1,3);
+initialArea=sampleWidth*sampleThickness;
+
+numElementsWidth=round(sampleWidth/pointSpacings(1));
+numElementsThickness=round(sampleThickness/pointSpacings(2));
+numElementsHeight=round(sampleHeight/pointSpacings(3));
+
+stretchLoad=0.7;
+displacementMagnitude=(stretchLoad*sampleHeight)-sampleHeight;
+
+%Material parameter set
+c1=1e-3; %Shear-modulus-like parameter
+m1=3; %Material parameter setting degree of non-linearity
+k_factor=1e2; %Bulk modulus factor 
+k=c1*k_factor; %Bulk modulus
+
+% FEA control settings
+numTimeSteps=10; %Number of time steps desired
+max_refs=25; %Max reforms
+max_ups=0; %Set to zero to use full-Newton iterations
+opt_iter=6; %Optimum number of iterations
+max_retries=5; %Maximum number of retires
+dtmin=(1/numTimeSteps)/100; %Minimum time step size
+dtmax=1/numTimeSteps; %Maximum time step size
+
+%% Creating model geometry and mesh
+
+%%
+% Create a box with hexahedral elements
+boxDim=[sampleWidth sampleThickness sampleHeight]; %Dimensions
+boxEl=[numElementsWidth numElementsThickness numElementsHeight]; %Number of elements
+outputStructType=2; %A structure compatible with mesh view
+[meshStruct]=hexMeshBox(boxDim,boxEl,outputStructType);
+
+E=meshStruct.elements;
+V=meshStruct.nodes;
+Fb=meshStruct.facesBoundary;
+Cb=meshStruct.boundaryMarker;
+
+X=V(:,1); Y=V(:,2); Z=V(:,3);
+VE=[mean(X(E),2) mean(Y(E),2) mean(Z(E),2)];
+
+elementMaterialIndices=ones(size(E,1),1);
+
+%%
+% Plotting model boundary surfaces
+
[…]
