### 2018-01-24 9deb5a2
Finalized first demo (DEMO_febio_0001_cube_uniaxial) using new febio_spec format and febioStruct2xml
