+++ docs/DEMO_febio_0001_cube_uniaxial.m
@@ -1,57 +1,72 @@
 %% DEMO_febio_0001_cube_uniaxial
 % Below is a demonstration for:
-% 1) Building an FEBio model for uniaxial loading
-% 2) Running the model
-% 3) Importing displacement and force results
+% * Building geometry for a cube with hexahedral elements
+% * Defining the boundary conditions 
+% * Coding the febio structure
+% * Running the model
+% * Importing and visualizing the displacement and stress results
+
+%% Keywords: 
+% * febio_spec version 2.5
+% * febio, FEBio
+% * uniaxial loading
+% * compression, tension, compressive, tensile
+% * displacement control, displacement boundary condition
+% * hexahedral elements, hex8
+% * cube, box, rectangular
+% * static, solid
+% * hyperelastic, Ogden
+% * Displacement logfile
+% * Stress logfile
 
 %%
 
 clear; close all; clc;
 
-%%
-% Plot settings
+%% Plot settings
 fontSize=20;
 faceAlpha1=0.8;
-faceAlpha2=1;
-edgeColor=0.25*ones(1,3);
-edgeWidth=1.5;
 markerSize=40;
 lineWidth=3;
 
-%%
-% Control parameters
+%% Control parameters
 
-% path names
+% Path names
 defaultFolder = fileparts(fileparts(mfilename('fullpath')));
 savePath=fullfile(defaultFolder,'data','temp');
 
 % Defining file names
-modelNameEnd='tempModel';
-modelName=fullfile(savePath,modelNameEnd);
-
-febioFebFileName=[modelName,'.feb']; %FEB file name
-febioLogFileName=[modelName,'.txt']; %FEBio log file name
-febioLogFileName_disp=[modelNameEnd,'_disp_out.txt']; %Log file name for exporting force
-febioLogFileName_force=[modelNameEnd,'_force_out.txt']; %Log file name for exporting force
-febioLogFileName_stress=[modelNameEnd,'_stress_out.txt']; %Log file name for exporting force
+febioFebFileNamePart='tempModel';
+febioFebFileName=fullfile(savePath,[febioFebFileNamePart,'.feb']); %FEB file name
+febioLogFileName=fullfile(savePath,[febioFebFileNamePart,'.txt']); %FEBio log file name
+febioLogFileName_disp=[febioFebFileNamePart,'_disp_out.txt']; %Log file name for exporting force
+febioLogFileName_force=[febioFebFileNamePart,'_force_out.txt']; %Log file name for exporting force
+febioLogFileName_stress=[febioFebFileNamePart,'_stress_out.txt']; %Log file name for exporting force
 
 %Specifying dimensions and number of elements
-sampleWidth=10;
-sampleThickness=10; 
-sampleHeight=10;
-pointSpacings=2*ones(1,3);
-initialArea=sampleWidth*sampleThickness;
+cubeSize=10; 
+sampleWidth=cubeSize; %Width 
+sampleThickness=cubeSize; %Thickness 
+sampleHeight=cubeSize; %Height
+pointSpacings=2*ones(1,3); %Desired point spacing between nodes
+numElementsWidth=round(sampleWidth/pointSpacings(1)); %Number of elemens in dir 1
+numElementsThickness=round(sampleThickness/pointSpacings(2)); %Number of elemens in dir 2
+numElementsHeight=round(sampleHeight/pointSpacings(3)); %Number of elemens in dir 3
 
-numElementsWidth=round(sampleWidth/pointSpacings(1));
-numElementsThickness=round(sampleThickness/pointSpacings(2));
-numElementsHeight=round(sampleHeight/pointSpacings(3));
[…]
