+++ data/temp/tempModel_stress_out.txt
@@ -2,1280 +2,1280 @@
 *Step  = 1
 *Time  = 0.1
 *Data  = sz
-1,-9.11959e-05
-2,-9.11959e-05
-3,-9.11959e-05
-4,-9.11959e-05
-5,-9.11959e-05
-6,-9.11959e-05
-7,-9.11959e-05
-8,-9.11959e-05
-9,-9.11959e-05
-10,-9.11959e-05
-11,-9.11959e-05
-12,-9.11959e-05
-13,-9.11959e-05
-14,-9.11959e-05
-15,-9.11959e-05
-16,-9.11959e-05
-17,-9.11959e-05
-18,-9.11959e-05
-19,-9.11959e-05
-20,-9.11959e-05
-21,-9.11959e-05
-22,-9.11959e-05
-23,-9.11959e-05
-24,-9.11959e-05
-25,-9.11959e-05
-26,-9.11959e-05
-27,-9.11959e-05
-28,-9.11959e-05
-29,-9.11959e-05
-30,-9.11959e-05
-31,-9.11959e-05
-32,-9.11959e-05
-33,-9.11959e-05
-34,-9.11959e-05
-35,-9.11959e-05
-36,-9.11959e-05
-37,-9.11959e-05
-38,-9.11959e-05
-39,-9.11959e-05
-40,-9.11959e-05
-41,-9.11959e-05
-42,-9.11959e-05
-43,-9.11959e-05
-44,-9.11959e-05
-45,-9.11959e-05
-46,-9.11959e-05
-47,-9.11959e-05
-48,-9.11959e-05
-49,-9.11959e-05
-50,-9.11959e-05
-51,-9.11959e-05
-52,-9.11959e-05
-53,-9.11959e-05
-54,-9.11959e-05
-55,-9.11959e-05
-56,-9.11959e-05
-57,-9.11959e-05
-58,-9.11959e-05
-59,-9.11959e-05
-60,-9.11959e-05
-61,-9.11959e-05
-62,-9.11959e-05
-63,-9.11959e-05
-64,-9.11959e-05
-65,-9.11959e-05
-66,-9.11959e-05
-67,-9.11959e-05
-68,-9.11959e-05
-69,-9.11959e-05
-70,-9.11959e-05
-71,-9.11959e-05
-72,-9.11959e-05
-73,-9.11959e-05
-74,-9.11959e-05
-75,-9.11959e-05
-76,-9.11959e-05
-77,-9.11959e-05
-78,-9.11959e-05
-79,-9.11959e-05
-80,-9.11959e-05
-81,-9.11959e-05
-82,-9.11959e-05
-83,-9.11959e-05
-84,-9.11959e-05
-85,-9.11959e-05
-86,-9.11959e-05
-87,-9.11959e-05
-88,-9.11959e-05
-89,-9.11959e-05
-90,-9.11959e-05
-91,-9.11959e-05
-92,-9.11959e-05
-93,-9.11959e-05
-94,-9.11959e-05
-95,-9.11959e-05
-96,-9.11959e-05
-97,-9.11959e-05
-98,-9.11959e-05
-99,-9.11959e-05
-100,-9.11959e-05
-101,-9.11959e-05
-102,-9.11959e-05
-103,-9.11959e-05
-104,-9.11959e-05
-105,-9.11959e-05
-106,-9.11959e-05
-107,-9.11959e-05
-108,-9.11959e-05
-109,-9.11959e-05
-110,-9.11959e-05
-111,-9.11959e-05
-112,-9.11959e-05
-113,-9.11959e-05
-114,-9.11959e-05
-115,-9.11959e-05
-116,-9.11959e-05
-117,-9.11959e-05
[…]

[…]
