### 2018-01-24 gibbonCode/GIBBON#17 @Kevin-Mattheus-Moerman
@wspr I did `git update-index --chmod=+x path/to/something' for the mac and linux version of tetGen. The latest version should have this. Can you check it out?
