### 2018-01-25 53b6ac5
Fixed order of field export in febioStruct2xml (important for febio)
