+++ docs/DEMO_febio_0051_cube_body_load.m
@@ -1,0 +1,369 @@
+%% DEMO_febio_0051_cube_body_load
+% Below is a demonstration for:
+% * Building geometry for a cube with hexahedral elements
+% * Defining the boundary conditions 
+% * Coding the febio structure
+% * Running the model
+% * Importing and visualizing the displacement and stress results
+
+%% Keywords: 
+% * febio_spec version 2.5
+% * febio, FEBio
+% * body load
+% * hexahedral elements, hex8
+% * cube, box, rectangular
+% * static, solid
+% * hyperelastic, Ogden
+% * Displacement logfile
+% * Force logfile
+
+%%
+
+clear; close all; clc;
+
+%% Plot settings
+fontSize=20;
+faceAlpha1=0.8;
+markerSize=40;
+lineWidth=3;
+
+%% Control parameters
+
+% Path names
+defaultFolder = fileparts(fileparts(mfilename('fullpath')));
+savePath=fullfile(defaultFolder,'data','temp');
+
+% Defining file names
+febioFebFileNamePart='tempModel';
+febioFebFileName=fullfile(savePath,[febioFebFileNamePart,'.feb']); %FEB file name
+febioLogFileName=fullfile(savePath,[febioFebFileNamePart,'.txt']); %FEBio log file name
+febioLogFileName_disp=[febioFebFileNamePart,'_disp_out.txt']; %Log file name for exporting force
+febioLogFileName_force=[febioFebFileNamePart,'_force_out.txt']; %Log file name for exporting force
+
+%Specifying dimensions and number of elements
+cubeSize=10; 
+sampleWidth=cubeSize; %Width 
+sampleThickness=cubeSize; %Thickness 
+sampleHeight=cubeSize; %Height
+pointSpacings=2*ones(1,3); %Desired point spacing between nodes
+numElementsWidth=round(sampleWidth/pointSpacings(1)); %Number of elemens in dir 1
+numElementsThickness=round(sampleThickness/pointSpacings(2)); %Number of elemens in dir 2
+numElementsHeight=round(sampleHeight/pointSpacings(3)); %Number of elemens in dir 3
+
+%Define applied displacement 
+bodyLoadType='const'; % or 'non-const'
+
+%Material parameter set
+c1=1e-3; %Shear-modulus-like parameter
+m1=8; %Material parameter setting degree of non-linearity
+k_factor=1e2; %Bulk modulus factor 
+k=c1*k_factor; %Bulk modulus
+
+% FEA control settings
+numTimeSteps=50; %Number of time steps desired
+max_refs=25; %Max reforms
+max_ups=0; %Set to zero to use full-Newton iterations
+opt_iter=6; %Optimum number of iterations
+max_retries=5; %Maximum number of retires
+dtmin=(1/numTimeSteps)/100; %Minimum time step size
+dtmax=1/numTimeSteps; %Maximum time step size
+
+%% Creating model geometry and mesh
+% A cube is created with tri-linear hexahedral (hex8) elements using the
+% |hexMeshBox| function. The function offers the boundary faces with
+% seperate labels for the top, bottom, left, right, front, and back sides
+% of the cube. As such these can be used to define boundary conditions on
+% the exterior. 
+
+% Create a box with hexahedral elements
+cubeDimensions=[sampleWidth sampleThickness sampleHeight]; %Dimensions
+cubeElementNumbers=[numElementsWidth numElementsThickness numElementsHeight]; %Number of elements
[…]
