+++ docs/DEMO_febio_0052_cube_discrete_springs.m
@@ -1,0 +1,381 @@
+%% DEMO_febio_0052_cube_discrete_springs
+% Below is a demonstration for: 
+% 
+% * Building geometry for a cube with hexahedral elements and discrete
+% elements on the edges
+% * Defining the boundary conditions 
+% * Coding the febio structure
+% * Running the model
+% * Importing and visualizing the displacement and stress results
+
+%% Keywords: 
+% * febio_spec version 2.5
+% * febio, FEBio
+% * discrete elements
+% * springs
+% * compression, tension, compressive, tensile
+% * displacement control, displacement boundary condition
+% * cube, box, rectangular
+% * static, solid
+% * hyperelastic, Ogden
+% * displacement logfile
+% * Stress logfile
+
+%%
+
+clear; close all; clc;
+
+%% Plot settings
+fontSize=20;
+faceAlpha1=0.8;
+markerSize=40;
+lineWidth=3;
+
+%% Control parameters
+
+% Path names
+defaultFolder = fileparts(fileparts(mfilename('fullpath')));
+savePath=fullfile(defaultFolder,'data','temp');
+
+% Defining file names
+febioFebFileNamePart='tempModel';
+febioFebFileName=fullfile(savePath,[febioFebFileNamePart,'.feb']); %FEB file name
+febioLogFileName=fullfile(savePath,[febioFebFileNamePart,'.txt']); %FEBio log file name
+febioLogFileName_disp=[febioFebFileNamePart,'_disp_out.txt']; %Log file name for exporting force
+febioLogFileName_force=[febioFebFileNamePart,'_force_out.txt']; %Log file name for exporting force
+
+%Specifying dimensions and number of elements
+cubeSize=10; 
+sampleWidth=cubeSize; %Width 
+sampleThickness=cubeSize; %Thickness 
+sampleHeight=cubeSize; %Height
+pointSpacings=1*ones(1,3); %Desired point spacing between nodes
+numElementsWidth=round(sampleWidth/pointSpacings(1)); %Number of elemens in dir 1
+numElementsThickness=round(sampleThickness/pointSpacings(2)); %Number of elemens in dir 2
+numElementsHeight=round(sampleHeight/pointSpacings(3)); %Number of elemens in dir 3
+
+%Define applied displacement 
+appliedStrain=0.2; %Linear strain (Only used to compute applied stretch)
+loadingOption='compression'; % or 'tension'
+switch loadingOption
+    case 'compression'
+        stretchLoad=1-appliedStrain; %The applied stretch for uniaxial loading
+    case 'tension'
+        stretchLoad=1+appliedStrain; %The applied stretch for uniaxial loading
+end
+displacementMagnitude=(stretchLoad*sampleHeight)-sampleHeight; %The displacement magnitude
+
+%Material parameter set
+c_spring=1e-3; %Spring "Youngs Modulus" 
+c1=c_spring/10000; %Solid shear-modulus-like parameter
+m1=2; %Solid material parameter setting degree of non-linearity
+k=c1; %Solid bulk modulus
+
+% FEA control settings
+numTimeSteps=10; %Number of time steps desired
+max_refs=25; %Max reforms
+max_ups=0; %Set to zero to use full-Newton iterations
+opt_iter=6; %Optimum number of iterations
+max_retries=5; %Maximum number of retires
+dtmin=(1/numTimeSteps)/100; %Minimum time step size
+dtmax=1/numTimeSteps; %Maximum time step size
+
[…]
