### 2018-01-31 c42af65
Added more demos for new febio_spec format
