+++ docs/DEMO_febio_0002_beam_force.m
@@ -1,21 +1,24 @@
-%% DEMO_febio_0051_cube_body_load
+%% DEMO_febio_0002_beam_force
 % Below is a demonstration for:
-% * Building geometry for a cube with hexahedral elements
+% 
+% * Building geometry for a beam with hexahedral elements
 % * Defining the boundary conditions 
 % * Coding the febio structure
 % * Running the model
-% * Importing and visualizing the displacement and stress results
+% * Importing and visualizing the displacement results
 
-%% Keywords: 
+%% Keywords
+%
 % * febio_spec version 2.5
 % * febio, FEBio
-% * body load
+% * beam force loading
+% * force control boundary condition
 % * hexahedral elements, hex8
-% * cube, box, rectangular
+% * beam, rectangular
 % * static, solid
 % * hyperelastic, Ogden
-% * Displacement logfile
-% * Force logfile
+% * displacement logfile
+% * stress logfile
 
 %%
 
@@ -41,17 +44,17 @@ febioLogFileName_disp=[febioFebFileNamePart,'_disp_out.txt']; %Log file name for
 febioLogFileName_force=[febioFebFileNamePart,'_force_out.txt']; %Log file name for exporting force
 
 %Specifying dimensions and number of elements
-cubeSize=10; 
-sampleWidth=cubeSize; %Width 
-sampleThickness=cubeSize; %Thickness 
-sampleHeight=cubeSize; %Height
+beamWidth=10; 
+sampleWidth=beamWidth; %Width 
+sampleThickness=4*beamWidth; %Thickness 
+sampleHeight=beamWidth; %Height
 pointSpacings=2*ones(1,3); %Desired point spacing between nodes
 numElementsWidth=round(sampleWidth/pointSpacings(1)); %Number of elemens in dir 1
 numElementsThickness=round(sampleThickness/pointSpacings(2)); %Number of elemens in dir 2
 numElementsHeight=round(sampleHeight/pointSpacings(3)); %Number of elemens in dir 3
 
-%Define applied displacement 
-bodyLoadType='const'; % or 'non-const'
+%Define applied force 
+appliedForce=[0 0 -5e-5]; 
 
 %Material parameter set
 c1=1e-3; %Shear-modulus-like parameter
@@ -60,7 +63,7 @@ k_factor=1e2; %Bulk modulus factor
 k=c1*k_factor; %Bulk modulus
 
 % FEA control settings
-numTimeSteps=50; %Number of time steps desired
+numTimeSteps=10; %Number of time steps desired
 max_refs=25; %Max reforms
 max_ups=0; %Set to zero to use full-Newton iterations
 opt_iter=6; %Optimum number of iterations
@@ -69,17 +72,17 @@ dtmin=(1/numTimeSteps)/100; %Minimum time step size
 dtmax=1/numTimeSteps; %Maximum time step size
 
 %% Creating model geometry and mesh
-% A cube is created with tri-linear hexahedral (hex8) elements using the
+% A beam is created with tri-linear hexahedral (hex8) elements using the
 % |hexMeshBox| function. The function offers the boundary faces with
 % seperate labels for the top, bottom, left, right, front, and back sides
 % of the cube. As such these can be used to define boundary conditions on
 % the exterior. 
 
 % Create a box with hexahedral elements
-cubeDimensions=[sampleWidth sampleThickness sampleHeight]; %Dimensions
[…]
