+++ docs/DEMO_febio_0006_sphere_indentation.m
@@ -1,7 +1,8 @@
-%% DEMO_febio_0002_beam_force
+%% DEMO_febio_0006_sphere_indentation
 % Below is a demonstration for:
 % 
-% * Building geometry for a beam with hexahedral elements
+% * Building geometry for a slab with hexahedral elements, and a
+% triangulated sphere. 
 % * Defining the boundary conditions 
 % * Coding the febio structure
 % * Running the model
@@ -11,10 +12,13 @@
 %
 % * febio_spec version 2.5
 % * febio, FEBio
-% * beam force loading
-% * force control boundary condition
+% * indentation
+% * contact, sliding, sticky, friction
+% * rigid body constraints
 % * hexahedral elements, hex8
-% * beam, rectangular
+% * triangular elements, tri3
+% * slab, block, rectangular
+% * sphere
 % * static, solid
 % * hyperelastic, Ogden
 % * displacement logfile
@@ -25,8 +29,9 @@
 clear; close all; clc;
 
 %% Plot settings
-fontSize=20;
+fontSize=15;
 faceAlpha1=0.8;
+faceAlpha2=0.3;
 markerSize=40;
 lineWidth=3;
 
@@ -43,18 +48,21 @@ febioLogFileName=fullfile(savePath,[febioFebFileNamePart,'.txt']); %FEBio log fi
 febioLogFileName_disp=[febioFebFileNamePart,'_disp_out.txt']; %Log file name for exporting force
 febioLogFileName_force=[febioFebFileNamePart,'_force_out.txt']; %Log file name for exporting force
 
-%Specifying dimensions and number of elements
-beamWidth=10; 
-sampleWidth=beamWidth; %Width 
-sampleThickness=4*beamWidth; %Thickness 
-sampleHeight=beamWidth; %Height
-pointSpacings=2*ones(1,3); %Desired point spacing between nodes
+%Specifying dimensions and number of elements for slab
+sampleHeight=5; %Height
+sampleWidth=sampleHeight*2; %Width 
+sampleThickness=sampleHeight*2; %Thickness 
+pointSpacings=0.5*ones(1,3); %Desired point spacing between nodes
 numElementsWidth=round(sampleWidth/pointSpacings(1)); %Number of elemens in dir 1
 numElementsThickness=round(sampleThickness/pointSpacings(2)); %Number of elemens in dir 2
 numElementsHeight=round(sampleHeight/pointSpacings(3)); %Number of elemens in dir 3
 
-%Define applied force 
-appliedForce=[0 0 -5e-5]; 
+%Sphere parameters
+numRefineStepsSphere=3; 
+sphereRadius=sampleHeight/2;
+
+%Define applied displacement
+sphereDisplacement=sphereRadius; 
 
 %Material parameter set
 c1=1e-3; %Shear-modulus-like parameter
@@ -66,11 +74,24 @@ k=c1*k_factor; %Bulk modulus
 numTimeSteps=10; %Number of time steps desired
 max_refs=25; %Max reforms
 max_ups=0; %Set to zero to use full-Newton iterations
-opt_iter=6; %Optimum number of iterations
+opt_iter=10; %Optimum number of iterations
 max_retries=5; %Maximum number of retires
 dtmin=(1/numTimeSteps)/100; %Minimum time step size
 dtmax=1/numTimeSteps; %Maximum time step size
 
+%Contact parameters
+contactInitialOffset=0.1;
+contactAlg=1;
+switch contactAlg
[…]
