+++ docs/DEMO_febio_0004_beam_twist.m
@@ -73,11 +73,10 @@ dtmin=(1/numTimeSteps)/100; %Minimum time step size
 dtmax=1/numTimeSteps; %Maximum time step size
 
 %% Creating model geometry and mesh
-% A beam is created with tri-linear hexahedral (hex8) elements using the
+% A box is created with tri-linear hexahedral (hex8) elements using the
 % |hexMeshBox| function. The function offers the boundary faces with
-% seperate labels for the top, bottom, left, right, front, and back sides
-% of the cube. As such these can be used to define boundary conditions on
-% the exterior. 
+% seperate labels for the top, bottom, left, right, front, and back sides.
+% As such these can be used to define boundary conditions on the exterior. 
 
 % Create a box with hexahedral elements
 beamDimensions=[sampleWidth sampleThickness sampleHeight]; %Dimensions
@@ -176,6 +175,12 @@ stepStruct.Control.time_stepper.opt_iter=opt_iter;
 stepStruct.Control.max_refs=max_refs;
 stepStruct.Control.max_ups=max_ups;
 
+%Add template based default settings to proposed control section
+[stepStruct.Control]=structComplete(stepStruct.Control,febio_spec.Control,1); %Complement provided with default if missing
+
+%Remove control field (part of template) since step specific control sections are used
+febio_spec=rmfield(febio_spec,'Control'); 
+
 %Material section
 febio_spec.Material.material{1}.ATTR.type='Ogden';
 febio_spec.Material.material{1}.ATTR.id=1;
@@ -194,7 +199,7 @@ febio_spec.Geometry.Nodes{1}.node.VAL=V; %The nodel coordinates
 % -> Elements
 febio_spec.Geometry.Elements{1}.ATTR.type='hex8'; %Element type of this set
 febio_spec.Geometry.Elements{1}.ATTR.mat=1; %material index for this set 
-febio_spec.Geometry.Elements{1}.ATTR.name='Cube'; %Name of the element set
+febio_spec.Geometry.Elements{1}.ATTR.name='Beam'; %Name of the element set
 febio_spec.Geometry.Elements{1}.elem.ATTR.id=(1:1:size(E,1))'; %Element id's
 febio_spec.Geometry.Elements{1}.elem.VAL=E;
 
