### 2018-02-02 def2b84
Fixed minor but in cFigure, added matlab based febioDoc and PDF file viewing
