### 2018-02-02 ce7ebad
Added gdoc to display or search GIBBON help and documentation
