+++ docs/DEMO_febio_0007_sphere_sliding.m
@@ -1,0 +1,532 @@
+%% DEMO_febio_0007_sphere_sliding
+% Below is a demonstration for:
+% 
+% * Building geometry for a slab with hexahedral elements, and a
+% triangulated sphere. 
+% * Defining the boundary conditions 
+% * Coding the febio structure
+% * Running the model
+% * Importing and visualizing the displacement results
+
+%% Keywords
+%
+% * febio_spec version 2.5
+% * febio, FEBio
+% * indentation
+% * contact, sliding, sticky, friction
+% * rigid body constraints
+% * hexahedral elements, hex8
+% * triangular elements, tri3
+% * slab, block, rectangular
+% * sphere
+% * static, solid
+% * hyperelastic, Ogden
+% * displacement logfile
+% * stress logfile
+
+%%
+
+clear; close all; clc;
+
+%% Plot settings
+fontSize=15;
+faceAlpha1=0.8;
+faceAlpha2=0.3;
+markerSize=40;
+lineWidth=3;
+
+%% Control parameters
+
+% Path names
+defaultFolder = fileparts(fileparts(mfilename('fullpath')));
+savePath=fullfile(defaultFolder,'data','temp');
+
+% Defining file names
+febioFebFileNamePart='tempModel';
+febioFebFileName=fullfile(savePath,[febioFebFileNamePart,'.feb']); %FEB file name
+febioLogFileName=fullfile(savePath,[febioFebFileNamePart,'.txt']); %FEBio log file name
+febioLogFileName_disp=[febioFebFileNamePart,'_disp_out.txt']; %Log file name for exporting force
+febioLogFileName_force=[febioFebFileNamePart,'_force_out.txt']; %Log file name for exporting force
+
+%Specifying dimensions and number of elements for slab
+sampleHeight=5; %Height
+sampleWidth=sampleHeight*4; %Width 
+sampleThickness=sampleHeight*1; %Thickness 
+pointSpacings=0.5*ones(1,3); %Desired point spacing between nodes
+numElementsWidth=round(sampleWidth/pointSpacings(1)); %Number of elemens in dir 1
+numElementsThickness=round(sampleThickness/pointSpacings(2)); %Number of elemens in dir 2
+numElementsHeight=round(sampleHeight/pointSpacings(3)); %Number of elemens in dir 3
+
+%Sphere parameters
+numRefineStepsSphere=3; 
+sphereRadius=sampleHeight/2;
+
+%Define applied displacement
+sphereIndentationDisplacement=sphereRadius/2; 
+sphereSlideDisplacement=sampleWidth-(sphereRadius*2); 
+
+%Material parameter set
+c1=1e-3; %Shear-modulus-like parameter
+m1=8; %Material parameter setting degree of non-linearity
+k_factor=1e2; %Bulk modulus factor 
+k=c1*k_factor; %Bulk modulus
+
+% FEA control settings
+numTimeSteps=10; %Number of time steps desired
+max_refs=25; %Max reforms
+max_ups=0; %Set to zero to use full-Newton iterations
+opt_iter=10; %Optimum number of iterations
+max_retries=5; %Maximum number of retires
+dtmin=(1/numTimeSteps)/100; %Minimum time step size
+dtmax=1/numTimeSteps; %Maximum time step size
+
+%Contact parameters
+contactInitialOffset=0.1;
+contactAlg=2;
+switch contactAlg
+    case 1
[…]
