### 2018-02-07 4325108
Added support for colormapped edge colors in gpatch
