### 2018-02-07 aa4ef42
Removed debugging statement from gpatch
