### 2018-02-07 a6707ab
Added simplexImIntersect for intersecting simplices
