### 2018-02-08 82ba1ad
Expanded simplexImIntersect o allow for labeling and use with multi-region simplices.
