+++ lib/simplexImIntersect.m
@@ -1,46 +1,56 @@
-function [logicOut]=simplexImIntersect(varargin)
+function [regionLabel]=simplexImIntersect(varargin)
 
-% [logicOut]=simplexImIntersect(F1,V1,V2,voxelSize)
+% [logicOut]=simplexImIntersect(F1,V1,C1,V2,voxelSize)
 % ------------------------------------------------------------------------
-% This function computes a logic which is true for vertices in V2 which are
-% deemed outside of the simplex defined by the faces array F1 and the
-% vertices array V1. Points are deemed outside based on the patch2Im
-% function, i.e. the following steps are used: 
-% 1) The simplex is converted to an image where voxels are 1 if they are in
-% or on the simplex and 0 if the are outside. 
+% This function uses the input simplex (or simplices) defined by F1
+% (faces), V1 (vertices), and C1 (boundary labels) to label the vertices V2
+% according to what simplex regions they are contained in. The labels are
+% stored in the output regionLabel. A vertex inside a simplex region is
+% labelled with the boundary label of the simplex region. Vertices outside
+% of all simplices are labelled with NaN. Points are labelled based on the
+% patch2Im function, i.e. the following steps are used: 
+% 1) The simplex is converted to an image where voxels intensities denote
+% simplex region labels. (see patch2Im).
 % 2) The vertices in V2 are converted to image coordinates in this image
 % 3) The image coordinates are converted to image voxel indices
-% 4) Voxel indices which point at voxels which are 0 in the simplex image
-% are outside. 
+% 4) Voxel indices are used to retrieve the labels from the simplex image. 
 %
-% The image constructed uses the optional input voxelSize (default if not
-% provided is half of the mean edge size of the input simplex). Although
-% all of the vertices where logicOut is 1 are out not all that are 0 in
-% logicOut are truely in. The level of accuracy of the logic depends on the
-% voxel size. If set too high then too many points are deemed outside. If
-% set too small then computational time is very large. 
+% The image constructed uses the optional input voxelSize (the default if
+% not provided is half of the mean edge size of the input simplex). The
+% completeness/accuracy of the labelling depends on the voxel size. Some
+% points are labelled as 0 which means they are found inside boundary
+% voxels and cannot, based on the current voxel size, be assigned as
+% outside or inside a particular region. All vertices labelled as NaN or a
+% value >0 are labelled correctly. However some of the vertices labelled as
+% 0 may actually be inside a simplex region or outside all simplex regions,
+% i.e. 0 denotes that their status is unknown given the voxel size used. 
+%
+% See also |patch2im|
 % 
 %
 % Kevin Mattheus Moerman
 % gibbon.toolbox@gmail.com
 % 
 % 2018/02/07:  Created
-%
+% 2018/02/08:  Expanded to allow for labelling and use with multi-region
+% simplices. 
 %------------------------------------------------------------------------
 
 %% Parse input 
 
 switch nargin    
-    case 3
-        F1=varargin{1};
-        V1=varargin{2};
-        V2=varargin{3};
-        voxelSize=[];
     case 4
         F1=varargin{1};
         V1=varargin{2};
-        V2=varargin{3};
-        voxelSize=varargin{4};
+        C1=varargin{3};
+        V2=varargin{4};
+        voxelSize=[];
+    case 5
+        F1=varargin{1};
[…]
