### 2018-02-08 1bacea0
Created GUI for installGibbon, added gitignore file to ignore help index files
