+++ installGibbon.m
@@ -1,62 +1,265 @@
 function installGibbon
 
-%% Add GIBBON paths
-disp('---INSTALLING GIBBON ---');
-disp('-> Adding gibbon paths');
-
+%% Add GIBBON library path so functions are known to use here
 gibbonPath=fileparts(mfilename('fullpath')); %Get the GIBBON path
 addpath(fullfile(gibbonPath,'lib')); %Add gibbon lib path so gibbon functions used here are known
 
+%% Settings
+
+W=40; %Scrollbar width
+
+%% Open figure
+
+figStruct.Name='Installing GIBBON';
+
+
+hf = cFigure(figStruct);
+hf.NumberTitle='off';
+
+%Remove tool and menu bars
+ht = findobj(allchild(hf),'flat','Type','uitoolbar');
+delete(ht);
+ht = findobj(allchild(hf),'flat','Type','uimenu');
+delete(ht);
+
+set(hf,'ResizeFcn',{@resizeAll,{hf}});
+
+% [left bottom width height]
+
+hf.UserData.W=W;
+
+%% Initilize text fields
+
+top_title='Installing GIBBON';
+hTextTitle = uicontrol(hf,'Style','text','String',top_title,...
+    'Position',[W hf.Position(4)-W hf.Position(3)-W*2 round(W/1.5)],...
+    'BackgroundColor',[1 1 1],'HorizontalAlignment','Left','FontSize',20,'FontWeight','normal');
+
+hf.UserData.uihandles.hTextTitle=hTextTitle;
+
+top_statement='Adding gibbon paths. Please wait...';
+hTextStatement = uicontrol(hf,'Style','text','String',top_statement,...
+    'Position',[W hf.Position(4)-2*W hf.Position(3)-W*2 round(W/1.5)],...
+    'BackgroundColor',[1 1 1],'HorizontalAlignment','Left','FontSize',16,'FontWeight','bold');
+
+hf.UserData.uihandles.hTextStatement=hTextStatement;
+
+%% Add image
+
+imagePath=fullfile(gibbonPath,'docs','img','GIBBON_overview.jpg');
+M=importdata(imagePath);
+
+figure(hf);
+image(M);
+axis tight; axis equal; axis off;
+hAxis=gca;
+
+hAxis.Units='pixels';
+% [left bottom width height]
+axisHeight=hf.Position(4)-7*W;
+axisWidth=hf.Position(3)-2*W;
+hAxis.Position=[round((hf.Position(3)-axisWidth)/2) round((hf.Position(4)-axisHeight))-6*W axisWidth axisHeight];
+
+hf.UserData.uihandles.hAxis=hAxis;
+
 %%
+drawnow;
+
+%% Adding paths
+
 [pathNames]=getSubPaths(gibbonPath);
 
-%%
-ignoreString='.'; %Target for ignoring .git and other hidden folders
+%Remove "hidden" folders with . in path name
+logicKeep=~contains(pathNames,'.');
+pathNames=pathNames(logicKeep);
+
+hw = waitbar(0,'Adding toolbox paths');
 for q=1:1:numel(pathNames)
-    pathNameNow=pathNames{q};
-    if ~any(strfind(pathNameNow,ignoreString))        
-        addpath(pathNameNow); %Add path
-    end    
+     pathNameNow=pathNames{q};     
+     addpath(pathNameNow); %Add path     
+     waitbar(q/numel(pathNames),hw);
 end
+close(hw)
 
-%% Add FEBio and export_fig paths
-disp('-> Adding 3rd party paths');
+hTextStatement.String='Done adding toolbox paths';
 
[…]
