### 2018-02-09 f79bf45
Emptied data/temp directory and will be added to gitignore
