+++ data/temp/iFEA_tempModel_strain_out.txt
@@ -1,2561 +1,0 @@
-*Title:
-*Step  = 1
-*Time  = 0.05
-*Data  = E1;E2;E3
-1, 0.0079012, 0.00732611, -0.0148875
-2, 0.0079012, 0.00732611, -0.0148875
-3, 0.0079012, 0.00732611, -0.0148875
-4, 0.0079012, 0.00732611, -0.0148875
-5, 0.0079012, 0.00732611, -0.0148875
-6, 0.0079012, 0.00732611, -0.0148875
-7, 0.0079012, 0.00732611, -0.0148875
-8, 0.0079012, 0.00732611, -0.0148875
-9, 0.0079012, 0.00732611, -0.0148875
-10, 0.0079012, 0.00732611, -0.0148875
-11, 0.0079012, 0.00732611, -0.0148875
-12, 0.0079012, 0.00732611, -0.0148875
-13, 0.0079012, 0.00732611, -0.0148875
-14, 0.0079012, 0.00732611, -0.0148875
-15, 0.0079012, 0.00732611, -0.0148875
-16, 0.0079012, 0.00732611, -0.0148875
-17, 0.0079012, 0.00732611, -0.0148875
-18, 0.0079012, 0.00732611, -0.0148875
-19, 0.0079012, 0.00732611, -0.0148875
-20, 0.0079012, 0.00732611, -0.0148875
-21, 0.0079012, 0.00732611, -0.0148875
-22, 0.0079012, 0.00732611, -0.0148875
-23, 0.0079012, 0.00732611, -0.0148875
-24, 0.0079012, 0.00732611, -0.0148875
-25, 0.0079012, 0.00732611, -0.0148875
-26, 0.0079012, 0.00732611, -0.0148875
-27, 0.0079012, 0.00732611, -0.0148875
-28, 0.0079012, 0.00732611, -0.0148875
-29, 0.0079012, 0.00732611, -0.0148875
-30, 0.0079012, 0.00732611, -0.0148875
-31, 0.0079012, 0.00732611, -0.0148875
-32, 0.0079012, 0.00732611, -0.0148875
-33, 0.0079012, 0.00732611, -0.0148875
-34, 0.0079012, 0.00732611, -0.0148875
-35, 0.0079012, 0.00732611, -0.0148875
-36, 0.0079012, 0.00732611, -0.0148875
-37, 0.0079012, 0.00732611, -0.0148875
-38, 0.0079012, 0.00732611, -0.0148875
-39, 0.0079012, 0.00732611, -0.0148875
-40, 0.0079012, 0.00732611, -0.0148875
-41, 0.0079012, 0.00732611, -0.0148875
-42, 0.0079012, 0.00732611, -0.0148875
-43, 0.0079012, 0.00732611, -0.0148875
-44, 0.0079012, 0.00732611, -0.0148875
-45, 0.0079012, 0.00732611, -0.0148875
-46, 0.0079012, 0.00732611, -0.0148875
-47, 0.0079012, 0.00732611, -0.0148875
-48, 0.0079012, 0.00732611, -0.0148875
-49, 0.0079012, 0.00732611, -0.0148875
-50, 0.0079012, 0.00732611, -0.0148875
-51, 0.0079012, 0.00732611, -0.0148875
-52, 0.0079012, 0.00732611, -0.0148875
-53, 0.0079012, 0.00732611, -0.0148875
-54, 0.0079012, 0.00732611, -0.0148875
-55, 0.0079012, 0.00732611, -0.0148875
-56, 0.0079012, 0.00732611, -0.0148875
-57, 0.0079012, 0.00732611, -0.0148875
[…]
