+++ data/temp/tempModel1.txt
@@ -1,1781 +1,0 @@
-===========================================================================
-         ________    _________   _________     __     _________            
-        |        |\ |        |\ |        |\   |  |\  /         \\          
-        |    ____|| |    ____|| |    __  ||   |__|| |    ___    ||         
-        |   |\___\| |   |\___\| |   |\_| ||    \_\| |   //  \   ||         
-        |   ||      |   ||      |   || | ||    __   |  ||    |  ||         
-        |   ||__    |   ||__    |   ||_| ||   |  |\ |  ||    |  ||         
-        |       |\  |       |\  |         \\  |  || |  ||    |  ||         
-        |    ___||  |    ___||  |    ___   || |  || |  ||    |  ||         
-        |   |\__\|  |   |\__\|  |   |\__|  || |  || |  ||    |  ||         
-        |   ||      |   ||      |   ||  |  || |  || |  ||    |  ||         
-        |   ||      |   ||___   |   ||__|  || |  || |   \\__/   ||         
-        |   ||      |        |\ |          || |  || |           ||         
-        |___||      |________|| |__________|| |__||  \_________//          
-                                                                           
-      F I N I T E   E L E M E N T S   F O R   B I O M E C H A N I C S      
-                                                                           
-                 --- v e r s i o n - 2 . 6 . 4 . 9579 ---                 
-                                                                           
-                                                                           
-  Musculoskeletal Research Laboratory                                      
-  University of Utah                                                       
-  http://febio.org                                                         
-                                                                           
-  FEBio is a registered trademark.                                         
-  copyright (c) 2006-2016 - All rights reserved                            
-                                                                           
- This is the NON-COMMERCIAL version of FEBio. This version may only be  
- used for non-commercial purposes as described in the license agreement.
- To obtain a valid commercial license file, please contact the developers.
-                                                                           
-===========================================================================
-
-
-
-
- FILES USED
-===========================================================================
-	Input file : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel1.feb
-	Plot file  : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel1.xplt
-	Log file   : /mnt/data/MATLAB/GIT/GIBBON/data/temp/tempModel1.txt
-
-
- MESH INFO
-===========================================================================
-	Number of materials ............................ : 2
-	Number of domains .............................. : 2
-	Number of nodes ................................ : 12100
-	Number of solid elements ....................... : 8394
-	Number of shell elements ....................... : 416
-
-
- CONTROL DATA
-===========================================================================
-	Module type .................................... : solid
-	Analysis type .................................. : quasi-static
-	Plane strain mode .............................. : no
-	Number of timesteps ............................ : 20
-	Time step size ................................. : 0.05
-	Auto time stepper activated .................... : yes
-	  Optimal nr of iterations ..................... : 5
-	  Minimum allowable step size .................. : 0.0005
-	  Maximum allowable step size .................. : 0.05
-	Number of loadcurves ........................... : 1
-	Number of displacement boundary conditions ..... : 0
-	Number of concentrated nodal forces ............ : 0
[…]

[…]
