### 2018-02-09 7215b60
Step 1 of moving data to zip files (removed dicoms)
