### 2018-02-09 99a90ec
Updated installGibbon to unzip zipped data upon installation
