### 2018-02-09 92623b7
Change font size on install gui
