### 2018-02-09 b7f0087
Ran testGibbon, updated license footer all code
