+++ paper/paper.md
@@ -17,7 +17,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-[GIBBON](https://github.com/gibbonCode/GIBBON) [@Moerman2017a], which loosely stands for Geometry and Image-Based Bioengineering add-ON, is an open-source MATLAB® toolbox providing a single open-source framework for many aspects of computational (bio)mechanics such as: image segmentation, meshing, boundary conditions specification, finite element analysis (FEA), and visualization. A schematic of the core functionality of GIBBON is shown in the figure below.
+[GIBBON](https://github.com/gibbonCode/GIBBON), which loosely stands for Geometry and Image-Based Bioengineering add-ON, is an open-source MATLAB® toolbox providing a single open-source framework for many aspects of computational (bio)mechanics such as: image segmentation, meshing, boundary conditions specification, finite element analysis (FEA), and visualization. A schematic of the core functionality of GIBBON is shown in the figure below.
 
 ![A Graphical summary of the GIBBON toolbox](GIBBON_overview.png)
 
@@ -30,9 +30,9 @@ Below is a more detailed discussion of the core functionality where reference is
 * **Lattice structures**: Various lattice structure tools have been implemented. One method to generate surface geometry for lattices is the use of triply-periodic functions (see `HELP_triplyPeriodicMinimal` and `DEMO_FEBio_trabeculae_compression`). Functions to convert element descriptions, such as tetrahedral and hexahedral elements, to lattice structures have also been implemented (`HELP_element2lattice` and `HELP_element2HexLattice`). These allow for the creation of 3D boundary conforming lattice structures on arbitrary input geometry. Exporting of hexahedral elements is also supported allowing for FEA on the created lattice structures.
 * **Finite element analysis (FEA)**: GIBBON interfaces with the free software [FEBio](http://febio.org/) [@Maas2012] for FEA (source code available on FEBio website). GIBBON can be used as a pre- and post- processor for FEA as it enables the code based development of meshes, boundary conditions, and input files. FEBio files can be directly exported based on dedicated MATLAB® structures. Furthermore, GIBBON can be used to start and control FEBio simulations. As such, iterative and inverse FEA (e.g. based on MATLAB® optimization routines) is also enabled. All `DEMO_FEBio_...` files contain FEBio examples, `DEMO_FEBio_iFEA_uniaxial_01` is a simple inverse FEA example.
 * **Visualization**  
-GIBBON expands the standard MATLAB® visualization capabilities by adding 3D image and voxel visualization (see `HELP_im2patch` and `HELP_sliceViewer`), meshed geometries (`HELP_gpatch`), finite element models (`HELP_element2patch`), and colormapped vector data (`HELP_quiverVec`), and all visualization methods enable multiple colormaps to be used in each figure or axis window. Furthermore GIBBON offers a custom figure window `cFigure` containing 3D rotation options (`HELP_vcw`) that mimic CAD behavior of 3D scene rendering, and high quality figure exporting options (`HELP_efw`). Advanced graphics animation creation and exporting capabilities through a figure window based GUI are also enabled (see `HELP_anim8`).
[…]
