+++ paper/paper.bib
@@ -33,15 +33,6 @@ volume = {134},
 year = {2012}
 }
 
-@misc{Moerman2017a,
-author = {Moerman, Kevin Mattheus},
-doi = {10.5281/ZENODO.888930},
-month = {sep},
-title = {{GIBBON: The Geometry and Image-Based Bioengineering add-On (Release: Hylobates albibarbis)}},
-url = {https://doi.org/10.5281/zenodo.888930{\#}.WbagG0xpIBY.mendeley},
-year = {2017}
-}
-
 @article{Sengeh2016,
 abstract = {{\textcopyright} 2016 Elsevier Ltd. Although the socket is critical in a prosthetic system for a person with limb amputation, the methods of its design are largely artisanal. A roadblock for a repeatable and quantitative socket design process is the lack of predictive and patient specific biomechanical models of the residuum. This study presents the evaluation of such a model using a combined experimental-numerical approach. The model geometry and tissue boundaries are derived from magnetic resonance imaging (MRI). The soft tissue non-linear elastic and viscoelastic mechanical behavior was evaluated using inverse finite element analysis (FEA) of in-vivo indentation experiments. A custom designed robotic in-vivo indentation system was used to provide a rich experimental data set of force versus time at 18 sites across a limb. During FEA, the tissues were represented by two layers, namely the skin-adipose layer and an underlying muscle-soft tissue complex. The non-linear elastic behavior was modeled using 2nd order Ogden hyperelastic formulations, and viscoelasticity was modeled using the quasi-linear theory of viscoelasticity. To determine the material parameters for each tissue, an inverse FEA based optimization routine was used that minimizes the combined mean of the squared force differences between the numerical and experimental force-time curves for indentations at 4 distinct anatomical regions on the residuum. The optimization provided the following material parameters for the skin-adipose layer: [. c=5.22kPa m=4.79 $\gamma$=3.57MPa $\tau$=0.32s] and for the muscle-soft tissue complex [. c=5.20kPam=4.78$\gamma$=3.47MPa$\tau$=0.34s]. These parameters were evaluated to predict the force-time curves for the remaining 14 anatomical locations. The mean percentage error (mean absolute error/ maximum experimental force) for these predictions was 7±3{\%}. The mean percentage error at the 4 sites used for the optimization was 4{\%}.},
 author = {Sengeh, D.M. and Moerman, K.M. and Petron, Arthur and Herr, Hugh M.},
@@ -100,6 +91,22 @@ title = {{Uniaxial and biaxial mechanical properties of porcine linea alba.}},
 url = {http://www.sciencedirect.com/science/article/pii/S1751616114003105},
 volume = {41},
 year = {2015}
+}
+
+@article{Moerman2016,
[…]
