+++ lib/dispStartTitleGibbonCode.m
@@ -14,7 +14,7 @@ disp(['--- ',msgText,' --- ',datestr(now)]); %Start message
 % image segmentation, image-based modeling, meshing, and finite element
 % analysis.
 % 
-% Copyright (C) 2017  Kevin Mattheus Moerman
+% Copyright (C) 2018  Kevin Mattheus Moerman
 % 
 % This program is free software: you can redistribute it and/or modify
 % it under the terms of the GNU General Public License as published by
