### 2018-02-09 d5e249d
Added DEMO_febio_0010_trabeculae_compression and updated paper
