+++ docs/DEMO_febio_0010_trabeculae_compression.m
@@ -1,8 +1,7 @@
-%% DEMO_febio_0052_cube_discrete_springs
+%% DEMO_febio_0010_trabeculae_compression
 % Below is a demonstration for: 
 % 
-% * Building geometry for a cube with hexahedral elements and discrete
-% elements on the edges
+% * Building geometry for trabecular structure with tetrahedral elements
 % * Defining the boundary conditions 
 % * Coding the febio structure
 % * Running the model
@@ -11,11 +10,11 @@
 %% Keywords: 
 % * febio_spec version 2.5
 % * febio, FEBio
-% * discrete elements
-% * springs
 % * compression, tension, compressive, tensile
 % * displacement control, displacement boundary condition
-% * cube, box, rectangular
+% * trabecular
+% * tetgen, meshing
+% * tetrahedral elements, tet4
 % * static, solid
 % * hyperelastic, Ogden
 % * displacement logfile
@@ -30,6 +29,7 @@ fontSize=20;
 faceAlpha1=0.8;
 markerSize=40;
 lineWidth=3;
+markerSize1=25; 
 
 %% Control parameters
 
@@ -44,15 +44,10 @@ febioLogFileName=fullfile(savePath,[febioFebFileNamePart,'.txt']); %FEBio log fi
 febioLogFileName_disp=[febioFebFileNamePart,'_disp_out.txt']; %Log file name for exporting force
 febioLogFileName_force=[febioFebFileNamePart,'_force_out.txt']; %Log file name for exporting force
 
-%Specifying dimensions and number of elements
-cubeSize=10; 
-sampleWidth=cubeSize; %Width 
-sampleThickness=cubeSize; %Thickness 
-sampleHeight=cubeSize; %Height
-pointSpacings=1*ones(1,3); %Desired point spacing between nodes
-numElementsWidth=round(sampleWidth/pointSpacings(1)); %Number of elemens in dir 1
-numElementsThickness=round(sampleThickness/pointSpacings(2)); %Number of elemens in dir 2
-numElementsHeight=round(sampleHeight/pointSpacings(3)); %Number of elemens in dir 3
+porousGeometryCase='g'; 
+ns=12; %Number of voxel steps across period for image data (roughly number of points on mesh period)
+nPeriods=[2 2 2]; %Number of periods in each direction
+sampleSize=10; %Heigh of the sample
 
 %Define applied displacement 
 appliedStrain=0.2; %Linear strain (Only used to compute applied stretch)
@@ -63,78 +58,184 @@ switch loadingOption
     case 'tension'
         stretchLoad=1+appliedStrain; %The applied stretch for uniaxial loading
 end
-displacementMagnitude=(stretchLoad*sampleHeight)-sampleHeight; %The displacement magnitude
+displacementMagnitude=(stretchLoad*sampleSize)-sampleSize; %The displacement magnitude
 
 %Material parameter set
-c_spring=1e-3; %Spring "Youngs Modulus" 
-c1=c_spring/100000; %Solid shear-modulus-like parameter
+c1=1e-3; %Solid shear-modulus-like parameter
 m1=2; %Solid material parameter setting degree of non-linearity
 k=c1; %Solid bulk modulus
 
 % FEA control settings
 numTimeSteps=10; %Number of time steps desired
-max_refs=25; %Max reforms
[…]

[…]
