### 2018-02-12 33ec984
Edits to paper file for JOSS and some associated demos
