+++ paper/paper.bib
@@ -135,3 +135,23 @@ title = {{Validation of continuously tagged MRI for the measurement of dynamic 3
 volume = {39},
 year = {2012}
 }
+
+@article{Lavecchia2018,
+abstract = {Low back pain is a major cause of disability and requires the development of new devices to treat pathologies and improve prognosis following surgery. Understanding the effects of new devices on the biomechanics of the spine is crucial in the development of new effective and functional devices. The aim of this study was to develop a preliminary parametric, scalable and anatomically accurate finite-element model of the lumbar spine allowing for the evaluation of the performance of spinal devices. The principal anatomical surfaces of the lumbar spine were first identified, and then accurately fitted from a previous model supplied by S14 Implants (Bordeaux, France). Finally, the reconstructed model was defined according to 17 parameters which are used to scale the model according to patient dimensions. The developed model, available as a toolbox named the lumbar model generator, enables generating a population of models using subject-specific dimensions obtained from data scans or averaged dimensions evaluated from the correlation analysis. This toolbox allows patient-specific assessment, taking into account individual morphological variation. The models have applications in the design process of new devices, evaluating the biomechanics of the spine and helping clinicians when deciding on treatment strategies.},
+author = {Lavecchia, C E and Espino, D M and Moerman, K M and Tse, K M and Robinson, D and Lee, P V S and Shepherd, D E T},
+doi = {10.1098/rsif.2017.0829},
+file = {:home/kevin/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Lavecchia et al. - 2018 - Lumbar model generator a tool for the automated generation of a parametric scalable model of the lumbar spine.pdf:pdf},
+issn = {1742-5662},
+journal = {Journal of the Royal Society, Interface},
+keywords = {biomechanics,finite-element analysis,lumbar spine,morphing,parametric model,spine},
+mendeley-groups = {paperBib},
+month = {jan},
+number = {138},
+pages = {20170829},
+pmid = {29298959},
+publisher = {The Royal Society},
+title = {{Lumbar model generator: a tool for the automated generation of a parametric scalable model of the lumbar spine.}},
+url = {http://www.ncbi.nlm.nih.gov/pubmed/29298959},
+volume = {15},
+year = {2018}
+}
