+++ paper/paper.md
@@ -1,10 +1,12 @@
 ---
 title: 'GIBBON: The Geometry and Image-Based Bioengineering add-On'
 tags:
+  - Computational mechanics
+  - Biomechanics
+  - Finite Element Analysis
   - Image-Based modeling
   - Visualization
   - Meshing
-  - Finite Element Analysis
 authors:
  - name: Kevin M Moerman
    orcid: 0000-0003-3768-4269
@@ -17,24 +19,23 @@ bibliography: paper.bib
 ---
 
 # Summary
-[GIBBON](https://github.com/gibbonCode/GIBBON), which loosely stands for Geometry and Image-Based Bioengineering add-ON, is an open-source MATLAB® toolbox providing a single open-source framework for many aspects of computational (bio)mechanics such as: image segmentation, meshing, boundary conditions specification, finite element analysis (FEA), and visualization. A schematic of the core functionality of GIBBON is shown in the figure below.
+[GIBBON](https://www.gibboncode.org), which loosely stands for Geometry and Image-Based Bioengineering add-ON, is a MATLAB® toolbox providing a single open-source framework for many aspects of computational (bio)mechanics such as: image segmentation, meshing, boundary conditions specification, finite element analysis (FEA), and visualization. A schematic of the core functionality of GIBBON is shown in the figure below.
 
 ![A Graphical summary of the GIBBON toolbox](GIBBON_overview.png)
 
 Below is a more detailed discussion of the core functionality where reference is made to implementations in the documentation.
 
-* **Image segmentation**: In patient-, or subject-specific biomechanics, the geometry information is often derived from image data (e.g. Magnetic Resonance Imaging (MRI)). GIBBON offers image filtering and smoothening methods, and has a graphical user interface for 3D image segmentation (see `HELP_imx.m`). The segmented image data can be converted to 3D surface models which can be meshed for FEA.
-* **Computer Aided Design (CAD) tools**: Sometimes geometry is instead imported or designed. Using GIBBON geometry can be imported from common mesh based CAD files (such as STL, see `HELP_import_STL`). However, several CAD style commands have also been implemented in GIBBON such as polygon rounding, revolution (see `HELP_polyRevolve`), extrusion (see `HELP_polyExtrude`), and sweeping and lofting (see `HELP_polyLoftLinear` and `HELP_sweepLoft`).
-* **Surface meshing tools**: 2D multi-region triangular meshing (see for instance `HELP_regionTriMesh2D` and `HELP_multiRegionTriMeshUneven2D`), resampling meshes geodesically (see `DEMO_geodesic_remeshing`), smoothening (`DEMO_surface_smooth_methods`), and surface mesh refinement (see `HELP_subtri`, `HELP_subTriDual` and `HELP_subQuad`), mesh type conversions (see `HELP_tri2quad`, `HELP_quad2tri`), and mesh dual computation (see `HELP_patch_dual`). Geometries can also be exported to the STL format e.g. for computer aided manufacture and 3D printing.
+* **Image segmentation**: In patient-, or subject-specific biomechanics, the geometry information is often derived from image data (e.g. Magnetic Resonance Imaging (MRI)). GIBBON offers image filtering and smoothening methods, and has a graphical user interface for 3D image segmentation (see `HELP_imx.m`). The segmented image data can be converted to 3D surface models (see `DEMO_imx_levelset_surface_compare`) which can be meshed for FEA (`HELP_runTetGen`).
[…]
