### 2018-02-12 gibbonCode/GIBBON#20 @Kevin-Mattheus-Moerman
Thanks! I addressed these points just now in 33ec984
