### 2018-02-12 e8b8743
Updated paper.bib to render references consistently
