+++ paper/paper.bib
@@ -17,7 +17,7 @@ year = {2015}
 
 @article{Maas2012,
 abstract = {In the field of computational biomechanics, investigators have primarily used commercial software that is neither geared toward biological applications nor sufficiently flexible to follow the latest developments in the field. This lack of a tailored software environment has hampered research progress, as well as dissemination of models and results. To address these issues, we developed the FEBio software suite (http://mrl.sci.utah.edu/software/febio), a nonlinear implicit finite element (FE) framework, designed specifically for analysis in computational solid biomechanics. This paper provides an overview of the theoretical basis of FEBio and its main features. FEBio offers modeling scenarios, constitutive models, and boundary conditions, which are relevant to numerous applications in biomechanics. The open-source FEBio software is written in C++, with particular attention to scalar and parallel performance on modern computer architectures. Software verification is a large part of the development and maintenance of FEBio, and to demonstrate the general approach, the description and results of several problems from the FEBio Verification Suite are presented and compared to analytical solutions or results from other established and verified FE codes. An additional simulation is described that illustrates the application of FEBio to a research problem in biomechanics. Together with the pre- and postprocessing software PREVIEW and POSTVIEW, FEBio provides a tailored solution for research and development in computational biomechanics.},
-author = {Maas, Steve A. and Ellis, Benjamin J. and Ateshian, Gerard A. and Weiss, Jeffrey A.},
+author = {Maas, S.A. and Ellis, B.J. and Ateshian, G.A. and Weiss, J.A.},
 doi = {10.1115/1.4005694},
 file = {:home/kevin/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Maas et al. - 2012 - FEBio Finite Elements for Biomechanics.pdf:pdf},
 isbn = {1528-8951 (Electronic)$\backslash$n0148-0731 (Linking)},
@@ -35,7 +35,7 @@ year = {2012}
 
 @article{Sengeh2016,
[…]
