### 2018-02-12 8f871b2
Edits to paper for JOSS
