### 2018-02-12 1db90e7
More minor edits to paper for JOSS and also readme
