### 2018-02-12 #506 @Kevin-Mattheus-Moerman
Sorry for spamming this thread. I had some minor paper edits and things to iron out.
