### 2018-02-12 #506 @Kevin-Mattheus-Moerman
@cMadan I believe I have addressed all of @siboles and @wspr comments.

@siboles and @wspr thanks for the thorough and useful review process! 🚀 🤖 🍰

FYI All functionality which was reviewed remains. I do want to report some new features which I'll highlight briefly.

* I upgraded GIBBON to also work with febio_spec 2.5 since the old functionality only covered febio_spec 2.0. New demos featuring febio_spec 2.5 are named `DEMO_febio_...`. The old demos using febio_spec 2.0 still work and remain named  `DEMO_FEBio...`.
* I implemented a new way to create the input file to allow for all of FEBio's functionality to be covered. (see `HELP_febioStruct2xml`). It is essentially a flexible way to handle XML files. So now any XML content can be coded with these tools allowing users to also code input files for experimental/novel FEBio features. This way GIBBON become ready not only for febio_spec 2.5 but also for any other future XML based release.
