### 2018-02-13 #506 @Kevin-Mattheus-Moerman
@cMadan Reviewer 2 (@wspr) commented about the fix for this issue that was raised https://github.com/gibbonCode/GIBBON/issues/17 it appears that is working now so I closed the issue.

I also implemented the suggested changes to the paper proposed in this issue https://github.com/gibbonCode/GIBBON/issues/20. I went ahead and closed the issue but perhaps @wspr can comment if my changes are deemed sufficient.
