+++ docs/DEMO_febio_0007_sphere_sliding.m
@@ -246,9 +246,14 @@ stepStruct.Control.max_ups=max_ups;
 %Add template based default settings to proposed control section
 [stepStruct.Control]=structComplete(stepStruct.Control,febio_spec.Control,1); %Complement provided with default if missing
 
+%Remove control field (part of template) since step specific control sections are used
+febio_spec=rmfield(febio_spec,'Control'); 
+
 %Step specific control section
 febio_spec.Step{1}.Control=stepStruct.Control;
+febio_spec.Step{1}.ATTR.id=1;
 febio_spec.Step{2}.Control=stepStruct.Control;
+febio_spec.Step{2}.ATTR.id=2;
     
 %Material section
 febio_spec.Material.material{1}.ATTR.type='Ogden';
