### 2018-02-18 63756d8
Updated febioStruct2xml to allow for high speed export, regenerated help and documentation
