### 2018-02-19 b90e5c4
Minor update to product page in help and documentation
