### 2018-02-19 #506 @Kevin-Mattheus-Moerman
@cmadan great! 🎉 Here is the DOI of the archived version:
[`https://doi.org/10.5281/zenodo.1175348`](https://doi.org/10.5281/zenodo.1175348)

@arfon the software version should be updated to be **v3.0.0**

@siboles @wspr thanks for the review! 🚀
